split `DirectedGraph` from `knit data lineage` implementation
to reuse it to implement `knit plan graph`.
