+++ cmd/knit/knitgraph/doc.go
@@ -1,0 +1,4 @@
+package knitgraph
+
+// knitgraph is a package for representing a graph (directed acyclic graph (DAG))
+// where the nodes are the Run, Data and Plan.
