+++ cmd/knit/subcommands/data/lineage/command.go
@@ -3,29 +3,26 @@ package lineage
 import (
 	"context"
 	"fmt"
-	"html"
-	"io"
 	"log"
-	"strconv"
-	"strings"
 
 	"github.com/opst/knitfab-api-types/data"
-	"github.com/opst/knitfab-api-types/misc/rfctime"
 	"github.com/opst/knitfab-api-types/runs"
 	"github.com/opst/knitfab-api-types/tags"
 	kenv "github.com/opst/knitfab/cmd/knit/env"
+	"github.com/opst/knitfab/cmd/knit/knitgraph"
 	"github.com/opst/knitfab/cmd/knit/subcommands/common"
 	"github.com/opst/knitfab/pkg/domain"
-	"github.com/opst/knitfab/pkg/utils/cmp"
+	"github.com/opst/knitfab/pkg/utils/args"
+	"github.com/opst/knitfab/pkg/utils/pointer"
 	"github.com/youta-t/flarc"
 
 	krst "github.com/opst/knitfab/cmd/knit/rest"
 )
 
 type Flag struct {
-	Upstream   bool   `flag:"upstream" alias:"u" help:"Trace the upstream of the specified Data."`
-	Downstream bool   `flag:"downstream" alias:"d" help:"Trace the downstream of the specified Data."`
-	Numbers    string `flag:"numbers" alias:"n" help:"Trace up to the specified depth. Trace to the upstream-most/downstream-most if 'all' is specified.,metavar=number of depth"`
+	Upstream   bool        `flag:"upstream" alias:"u" help:"Trace the upstream of the specified Data."`
+	Downstream bool        `flag:"downstream" alias:"d" help:"Trace the downstream of the specified Data."`
+	Numbers    *args.Depth `flag:"numbers" alias:"n" help:"Trace up to the specified depth. Trace to the upstream-most/downstream-most if 'all' is specified.,metavar=number of depth"`
 }
 
 type Option struct {
@@ -40,10 +37,10 @@ type Traverser struct {
 type Runner func(
 	ctx context.Context,
 	client krst.KnitClient,
-	graph *DirectedGraph,
+	graph *knitgraph.DirectedGraph,
 	knitId string,
-	depth int,
-) (*DirectedGraph, error)
+	depth args.Depth,
+) (*knitgraph.DirectedGraph, error)
 
 var _ Runner = TraceUpStream
 var _ Runner = TraceDownStream
@@ -75,7 +72,7 @@ func New(
 		Flag{
 			Upstream:   false,
 			Downstream: false,
-			Numbers:    "3",
+			Numbers:    pointer.Ref(args.NewDepth(3)),
 		},
 		flarc.Args{
 			{
@@ -141,25 +138,13 @@ func Task(option Traverser) common.Task[Flag] {
 	) error {
 		flags := cl.Flags()
 		numbers := flags.Numbers
-		var depth int
-		var err error
-
-		if numbers == "all" {
-			depth = -1
-		} else {
-			depth, err = strconv.Atoi(numbers)
-			if err != nil {
-				return fmt.Errorf("%w: invalid number %s", flarc.ErrUsage, numbers)
-			}
-			if depth <= 0 {
-				return fmt.Errorf("%w: depth should be natural number", flarc.ErrUsage)
-			}
+		if !numbers.IsInfinity() && numbers.Value() == 0 {
+			return fmt.Errorf("%w: --numbers must be a positive integer or 'all'", flarc.ErrUsage)
 		}
-
 		shoudUpstream := flags.Upstream
 		shoudDownstream := flags.Downstream
 		knitId := cl.Args()[ARG_KNITID][0]
-		graph := NewDirectedGraph()
+		graph := knitgraph.NewDirectedGraph()
 
 		// If both flags are not set, upstream and downstream will be traced.
 		if !shoudUpstream && !shoudDownstream {
@@ -168,16 +153,18 @@ func Task(option Traverser) common.Task[Flag] {
 		}
 
 		if shoudUpstream {
-			graph, err = option.ForUpstream(ctx, client, graph, knitId, depth)
+			_graph, err := option.ForUpstream(ctx, client, graph, knitId, *numbers)
 			if err != nil {
 				return err
 			}
+			graph = _graph
 		}
 		if shoudDownstream {
-			graph, err = option.ForDownstream(ctx, client, graph, knitId, depth)
+			_graph, err := option.ForDownstream(ctx, client, graph, knitId, *numbers)
 			if err != nil {
 				return err
 			}
+			graph = _graph
 		}
 
 		if err := graph.GenerateDot(cl.Stdout(), knitId); err != nil {
@@ -186,317 +173,6 @@ func Task(option Traverser) common.Task[Flag] {
 		logger.Println("success to output dot format")
 		return nil
 	}
-}
-
-type DirectedGraph struct {
-	DataNodes     map[string]DataNode
-	RunNodes      map[string]RunNode
-	RootNodes     []string          //to runId
-	EdgesFromRun  map[string][]Edge //key:from runId, value:to knitId
-	EdgesFromData map[string][]Edge //key:from knitId, value:to runId
-	//The order of these array becomes the order of description when outputting to the dot format.
-	KeysDataNode []string
-	KeysRunNode  []string
-}
-
-func NewDirectedGraph() *DirectedGraph {
-	return &DirectedGraph{
-		DataNodes:     map[string]DataNode{},
-		RunNodes:      map[string]RunNode{},
-		RootNodes:     []string{},
-		EdgesFromRun:  map[string][]Edge{},
-		EdgesFromData: map[string][]Edge{},
-		KeysDataNode:  make([]string, 0),
-		KeysRunNode:   make([]string, 0),
-	}
-}
-
-type DataNode struct {
-	KnitId string
-	Tags   []tags.Tag
-	// FromRunId is coreresponding runId in Upstream of apidata.Detail.
-	// ToRunIds is coreresponding runIds in Downstreams of apidata.Detail.
-	// The reason for holding these two types of runIds in this node is related to the data tracking algorithm.
-	// According to the algorithm's specifications, when data is first identified,
-	// there may be runIds contained in that data that are not yet determined to be added to the graph.
-	// The timing for confirming the addition is when the algorithm identifies that data again.
-	FromRunId string
-	ToRunIds  []string
-}
-
-func (d *DataNode) Equal(o *DataNode) bool {
-	return d.KnitId == o.KnitId &&
-		cmp.SliceContentEqWith(d.Tags, o.Tags, tags.Tag.Equal)
-}
-
-func (d *DataNode) ToDot(w io.Writer, isArgKnitId bool) error {
-	knitId := d.KnitId
-
-	systemtag := []string{}
-	userTags := []string{}
-	for _, tag := range d.Tags {
-		switch tag.Key {
-		case domain.KeyKnitTimestamp:
-			tsp, err := rfctime.ParseRFC3339DateTime(tag.Value)
-			if err != nil {
-				return err
-			}
-			systemtag = append(
-				systemtag,
-				html.EscapeString(tsp.Time().Local().Format(rfctime.RFC3339DateTimeFormat)),
-			)
-			continue
-		case domain.KeyKnitTransient:
-			systemtag = append(systemtag, html.EscapeString(tag.String()))
-			continue
-		case tags.KeyKnitId:
-			continue
-		}
-
-		userTags = append(
-			userTags,
-			fmt.Sprintf(`<B>%s</B>:%s`, html.EscapeString(tag.Key), html.EscapeString(tag.Value)),
-		)
-	}
-	subheader := ""
-	if len(systemtag) != 0 {
-		subheader = fmt.Sprintf(
-			`<TR><TD COLSPAN="2"><FONT POINT-SIZE="8">%s</FONT></TD></TR>`,
-			strings.Join(systemtag, " | "),
-		)
-	}
-
-	//The background color of the data node that is the argument gets highlighted from the others.
-	bgColor := map[bool]string{true: "#d4ecc6", false: "#FFFFFF"}[isArgKnitId]
-	_, err := fmt.Fprintf(
-		w,
-		`	"d%s"[
-		shape=none
-		color="#1c9930"
-		label=<
-			<TABLE CELLSPACING="0">
-				<TR><TD BGCOLOR="#1c9930"><FONT COLOR="#FFFFFF"><B>Data</B></FONT></TD><TD BGCOLOR="%s">knit#id: %s</TD></TR>
-				%s
-				<TR><TD COLSPAN="2">%s</TD></TR>
-			</TABLE>
-		>
-	];
-`,
-		knitId,
-		bgColor,
-		html.EscapeString(knitId),
-		subheader,
-		strings.Join(userTags, "<BR/>"),
-		// "d" is a prefix used to denote a data node in dot format.
-	)
-	return err
-
-}
-
-type RunNode struct {
-	runs.Summary
-}
-
-func (r *RunNode) ToDot(w io.Writer) error {
-	title := ""
-	if r.Plan.Image != nil {
-		title = "image = " + r.Plan.Image.String()
-	} else if r.Plan.Name != "" {
-		title = r.Plan.Name
-	}
-
-	status := html.EscapeString(r.Status)
-	switch domain.KnitRunStatus(r.Status) {
-	case domain.Deactivated:
-		status = fmt.Sprintf(`<FONT COLOR="gray"><B>%s</B></FONT>`, status)
-	case domain.Completing, domain.Done:
-		status = fmt.Sprintf(`<FONT COLOR="#007700"><B>%s</B></FONT>`, status)
-	case domain.Aborting, domain.Failed:
-		status = fmt.Sprintf(`<FONT COLOR="red"><B>%s</B></FONT>`, status)
-	}
-
-	_, err := fmt.Fprintf(
-		w,
-		`	"r%s"[
-		shape=none
-		color=orange
-		label=<
-			<TABLE CELLSPACING="0">
-				<TR><TD BGCOLOR="orange"><FONT COLOR="#FFFFFF"><B>Run</B></FONT></TD><TD>%s</TD><TD>id: %s</TD></TR>
-				<TR><TD COLSPAN="3"><FONT POINT-SIZE="8">last updated: %s</FONT></TD></TR>
-				<TR><TD COLSPAN="3">%s</TD></TR>
-			</TABLE>
-		>
-	];
-`,
-		r.RunId,
-		status,
-		html.EscapeString(r.RunId),
-		html.EscapeString(r.UpdatedAt.Time().Local().Format(rfctime.RFC3339DateTimeFormat)),
-		html.EscapeString(title),
-		// "r" is a prefix used to denote a run node in dot format.
-	)
-	return err
-}
-
-type Edge struct {
-	ToId  string
-	Label string //mountpath
-}
-
-func (g *DirectedGraph) AddDataNode(data data.Detail) {
-	g.DataNodes[data.KnitId] = DataNode{
-		KnitId:    data.KnitId,
-		Tags:      data.Tags,
-		FromRunId: data.Upstream.Run.RunId,
-		ToRunIds: func() []string {
-			runIds := []string{}
-			for _, ds := range data.Downstreams {
-				runIds = append(runIds, ds.Run.RunId)
-			}
-			return runIds
-		}(),
-	}
-	g.KeysDataNode = append(g.KeysDataNode, data.KnitId)
-}
-
-func (g *DirectedGraph) AddRunNode(run runs.Detail) {
-	g.RunNodes[run.RunId] = RunNode{
-		Summary: run.Summary,
-	}
-	g.KeysRunNode = append(g.KeysRunNode, run.RunId)
-}
-
-// This method assumes that the run nodes of the edge to be added are included in the graph.
-func (g *DirectedGraph) AddRootNode(runId string) {
-	g.RootNodes = append(g.RootNodes, runId)
-}
-
-// This method assumes that the nodes of the edge to be added are included in the graph.
-func (g *DirectedGraph) AddEdgeFromRun(runId string, knitId string, label string) {
-	g.EdgesFromRun[runId] = append(g.EdgesFromRun[runId], Edge{ToId: knitId, Label: label})
-}
-
-// This method assumes that the nodes of the edge to be added are included in the graph.
-func (g *DirectedGraph) AddEdgeFromData(knitId string, runId string, label string) {
-	g.EdgesFromData[knitId] = append(g.EdgesFromData[knitId], Edge{ToId: runId, Label: label})
-}
-
-func (g *DirectedGraph) GenerateDotFromNodes(w io.Writer, argKnitId string) error {
-	for _, knitId := range g.KeysDataNode {
-		if data, ok := g.DataNodes[knitId]; ok {
-			if err := data.ToDot(w, argKnitId == knitId); err != nil {
-				return err
-			}
-		}
-	}
-	for _, runId := range g.KeysRunNode {
-		if run, ok := g.RunNodes[runId]; ok {
-			if err := run.ToDot(w); err != nil {
-				return err
-			}
-		}
-	}
-	for i := range g.RootNodes {
-		_, err := fmt.Fprintf(
-			w,
-			`	"root#%d"[shape=Mdiamond];
-`,
-			i)
-		if err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func (g DirectedGraph) GenerateDotFromEdges(w io.Writer) error {
-	err := writeEdgesToDot(w, g.KeysDataNode, g.EdgesFromData, "d", "r")
-	if err != nil {
-		return err
-	}
-
-	err = writeEdgesToDot(w, g.KeysRunNode, g.EdgesFromRun, "r", "d")
-	if err != nil {
-		return err
-	}
-
-	for i, runId := range g.RootNodes {
-		_, err := fmt.Fprintf(
-			w,
-			`	"root#%d" -> "r%s";
-`,
-			i, runId,
-		)
-		if err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func writeEdgesToDot(
-	w io.Writer,
-	keys []string,
-	edgesMap map[string][]Edge,
-	fromPrefix, toPrefix string) error {
-	for _, id := range keys {
-		if edges, ok := edgesMap[id]; ok {
-			for _, edge := range edges {
-				toId := edge.ToId
-				_, err := fmt.Fprintf(
-					w,
-					`	"%s%s" -> "%s%s" [label="%s"];
-`,
-					fromPrefix, id,
-					toPrefix, toId,
-					edge.Label,
-				)
-				if err != nil {
-					return err
-				}
-			}
-		}
-	}
-	return nil
-}
-
-func (g *DirectedGraph) GenerateDot(w io.Writer, argKnitId string) error {
-	_, err := w.Write([]byte(`digraph G {
-	node [shape=record fontsize=10]
-	edge [fontsize=10]
-
-`))
-	if err != nil {
-		return err
-	}
-
-	err = g.GenerateDotFromNodes(w, argKnitId)
-	if err != nil {
-		return err
-	}
-	_, err = w.Write([]byte("\n"))
-	if err != nil {
-		return err
-	}
-
-	err = g.GenerateDotFromEdges(w)
-	if err != nil {
-		return err
-	}
-	_, err = w.Write([]byte("\n}"))
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func ErrFindDataWithKnitId(knitId string, err error) error {
-	return fmt.Errorf("%w: during searching data %s", err, knitId)
-}
-
-func ErrGetRunWithRunId(runId string, err error) error {
-	return fmt.Errorf("%w: during searching run %s", err, runId)
 }
 
 var ErrNotFoundData = fmt.Errorf("data not found")
@@ -508,7 +184,7 @@ func errNotFoundData(knitId string) error {
 func AddEdgeFromRunToLog(
 	ctx context.Context,
 	client krst.KnitClient,
-	graph *DirectedGraph,
+	graph *knitgraph.DirectedGraph,
 	run runs.Detail,
 ) (string, error) {
 
@@ -537,15 +213,15 @@ func AddEdgeFromRunToLog(
 func TraceDownStream(
 	ctx context.Context,
 	client krst.KnitClient,
-	graph *DirectedGraph,
+	graph *knitgraph.DirectedGraph,
 	rootKnitId string,
-	maxDepth int,
-) (*DirectedGraph, error) {
+	maxDepth args.Depth,
+) (*knitgraph.DirectedGraph, error) {
 
 	startKnitIds := []string{rootKnitId}
-	depth := 0
+	depth := uint(0)
 
-	for (maxDepth == -1 || depth < maxDepth) && 0 < len(startKnitIds) {
+	for (maxDepth.IsInfinity() || depth < maxDepth.Value()) && 0 < len(startKnitIds) {
 		var err error
 		graph, startKnitIds, err = TraceDownstreamOneStep(ctx, client, graph, startKnitIds)
 		if err != nil {
@@ -560,9 +236,9 @@ func TraceDownStream(
 func TraceDownstreamOneStep(
 	ctx context.Context,
 	client krst.KnitClient,
-	graph *DirectedGraph,
+	graph *knitgraph.DirectedGraph,
 	knitIds []string,
-) (*DirectedGraph, []string, error) {
+) (*knitgraph.DirectedGraph, []string, error) {
 
 	TotalnextKnitIds := []string{}
 
@@ -582,9 +258,9 @@ func TraceDownstreamOneStep(
 func TraceDownstreamForSingleNode(
 	ctx context.Context,
 	client krst.KnitClient,
-	graph *DirectedGraph,
+	graph *knitgraph.DirectedGraph,
 	knitId string,
-) (*DirectedGraph, []string, error) {
+) (*knitgraph.DirectedGraph, []string, error) {
 
 	//1. If the argument's data does not exist in the graph, add that data　to the graph.
 	if _, ok := graph.DataNodes[knitId]; !ok {
@@ -604,7 +280,7 @@ func TraceDownstreamForSingleNode(
 		//If the run does not exist in the graph, add it to the graph.
 		run, err := client.GetRun(ctx, toRunId)
 		if err != nil {
-			return graph, []string{}, ErrGetRunWithRunId(toRunId, err)
+			return graph, []string{}, knitgraph.ErrGetRunWithRunId(toRunId, err)
 		}
 		graph.AddRunNode(run)
 
@@ -659,15 +335,15 @@ func TraceDownstreamForSingleNode(
 func TraceUpStream(
 	ctx context.Context,
 	client krst.KnitClient,
-	graph *DirectedGraph,
+	graph *knitgraph.DirectedGraph,
 	rootKnitId string,
-	maxDepth int,
-) (*DirectedGraph, error) {
+	maxDepth args.Depth,
+) (*knitgraph.DirectedGraph, error) {
 
 	startKnitIds := []string{rootKnitId}
-	depth := 0
+	depth := uint(0)
 
-	for (maxDepth == -1 || depth < maxDepth) && 0 < len(startKnitIds) {
+	for (maxDepth.IsInfinity() || depth < maxDepth.Value()) && 0 < len(startKnitIds) {
 		var err error
 		graph, startKnitIds, err = TraceUpstreamOneStep(ctx, client, graph, startKnitIds)
 		if err != nil {
@@ -682,9 +358,9 @@ func TraceUpStream(
 func TraceUpstreamOneStep(
 	ctx context.Context,
 	client krst.KnitClient,
-	graph *DirectedGraph,
+	graph *knitgraph.DirectedGraph,
 	knitIds []string,
-) (*DirectedGraph, []string, error) {
+) (*knitgraph.DirectedGraph, []string, error) {
 
 	TotalnextKnitIds := []string{}
 
@@ -704,9 +380,9 @@ func TraceUpstreamOneStep(
 func TraceUpstreamForSingleNode(
 	ctx context.Context,
 	client krst.KnitClient,
-	graph *DirectedGraph,
+	graph *knitgraph.DirectedGraph,
 	knitId string,
-) (*DirectedGraph, []string, error) {
+) (*knitgraph.DirectedGraph, []string, error) {
 
 	///1. If the argument's data does not exist in the graph, add that data　to the graph.
 	if _, ok := graph.DataNodes[knitId]; !ok {
@@ -725,7 +401,7 @@ func TraceUpstreamForSingleNode(
 		//If the run does not exist in the graph, add it to the graph.
 		run, err := client.GetRun(ctx, fromRunId)
 		if err != nil {
-			return graph, []string{}, ErrGetRunWithRunId(fromRunId, err)
+			return graph, []string{}, knitgraph.ErrGetRunWithRunId(fromRunId, err)
 		}
 		graph.AddRunNode(run)
 
@@ -780,7 +456,7 @@ func TraceUpstreamForSingleNode(
 func getData(ctx context.Context, client krst.KnitClient, knitId string) (data.Detail, error) {
 	datas, err := client.FindData(ctx, []tags.Tag{knitIdTag(knitId)}, nil, nil)
 	if err != nil {
-		return data.Detail{}, ErrFindDataWithKnitId(knitId, err)
+		return data.Detail{}, knitgraph.ErrFindDataWithKnitId(knitId, err)
 	}
 	if len(datas) == 0 {
 		return data.Detail{}, errNotFoundData(knitId)
