+++ cmd/knit/subcommands/data/lineage/command_test.go
@@ -9,6 +9,7 @@ import (
 	"time"
 
 	kprof "github.com/opst/knitfab/cmd/knit/config/profiles"
+	"github.com/opst/knitfab/cmd/knit/knitgraph"
 	"github.com/youta-t/flarc"
 
 	"github.com/opst/knitfab-api-types/data"
@@ -22,13 +23,15 @@ import (
 	"github.com/opst/knitfab/cmd/knit/subcommands/internal/commandline"
 	"github.com/opst/knitfab/cmd/knit/subcommands/logger"
 	"github.com/opst/knitfab/pkg/domain"
+	"github.com/opst/knitfab/pkg/utils/args"
 	"github.com/opst/knitfab/pkg/utils/cmp"
+	"github.com/opst/knitfab/pkg/utils/pointer"
 	"github.com/opst/knitfab/pkg/utils/try"
 )
 
 func TestLineageCommand(t *testing.T) {
-	mockGraphFromUpstream := lineage.NewDirectedGraph()
-	mockGraphFromDownstream := lineage.NewDirectedGraph()
+	mockGraphFromUpstream := knitgraph.NewDirectedGraph()
+	mockGraphFromDownstream := knitgraph.NewDirectedGraph()
 
 	type when struct {
 		knitId string
@@ -48,13 +51,10 @@ func TestLineageCommand(t *testing.T) {
 			funcForUpstream := func(
 				ctx context.Context,
 				client krst.KnitClient,
-				graph *lineage.DirectedGraph,
+				graph *knitgraph.DirectedGraph,
 				knitId string,
-				depth int,
-			) (*lineage.DirectedGraph, error) {
-				if when.flag.Numbers == "all" && depth != -1 {
-					t.Errorf("should call task for traceupstream with depth -1")
-				}
+				depth args.Depth,
+			) (*knitgraph.DirectedGraph, error) {
 				if !when.flag.Upstream && when.flag.Downstream {
 					t.Errorf("should not call task for traceupstream")
 				}
@@ -63,13 +63,10 @@ func TestLineageCommand(t *testing.T) {
 			funcForDownstream := func(
 				ctx context.Context,
 				client krst.KnitClient,
-				graph *lineage.DirectedGraph,
+				graph *knitgraph.DirectedGraph,
 				knitId string,
-				depth int,
-			) (*lineage.DirectedGraph, error) {
-				if when.flag.Numbers == "all" && depth != -1 {
-					t.Errorf("should call task for tracedownstream with depth -1")
-				}
+				depth args.Depth,
+			) (*knitgraph.DirectedGraph, error) {
 				if when.flag.Upstream && !when.flag.Downstream {
 					t.Errorf("should not call task for tracedownstream")
 				}
@@ -121,7 +118,7 @@ func TestLineageCommand(t *testing.T) {
 			flag: lineage.Flag{
 				Upstream:   true,
 				Downstream: false,
-				Numbers:    "3",
+				Numbers:    pointer.Ref(args.NewDepth(3)),
 			},
 			err: nil,
 		},
@@ -136,7 +133,7 @@ func TestLineageCommand(t *testing.T) {
 			flag: lineage.Flag{
 				Upstream:   false,
 				Downstream: true,
-				Numbers:    "3",
+				Numbers:    pointer.Ref(args.NewDepth(3)),
 			},
 			err: nil,
 		},
@@ -151,7 +148,7 @@ func TestLineageCommand(t *testing.T) {
 			flag: lineage.Flag{
 				Upstream:   true,
 				Downstream: true,
-				Numbers:    "3",
+				Numbers:    pointer.Ref(args.NewDepth(3)),
 			},
 			err: nil,
 		},
@@ -166,7 +163,7 @@ func TestLineageCommand(t *testing.T) {
 			flag: lineage.Flag{
 				Upstream:   false,
 				Downstream: false,
-				Numbers:    "3",
+				Numbers:    pointer.Ref(args.NewDepth(3)),
 			},
 			err: nil,
 		},
@@ -181,7 +178,7 @@ func TestLineageCommand(t *testing.T) {
 			flag: lineage.Flag{
 				Upstream:   false,
 				Downstream: false,
-				Numbers:    "1",
+				Numbers:    pointer.Ref(args.NewDepth(1)),
 			},
 			err: nil,
 		},
@@ -196,7 +193,7 @@ func TestLineageCommand(t *testing.T) {
 			flag: lineage.Flag{
 				Upstream:   false,
 				Downstream: false,
-				Numbers:    "all",
+				Numbers:    pointer.Ref(args.NewInfinityDepth()),
 			},
 			err: nil,
 		},
@@ -211,22 +208,7 @@ func TestLineageCommand(t *testing.T) {
 			flag: lineage.Flag{
 				Upstream:   false,
 				Downstream: false,
-				Numbers:    "0",
-			},
-			err: nil,
-		},
-		then{
-			err: flarc.ErrUsage,
-		},
-	))
-
-	t.Run("when number of depth flag is letter except `all`, it shoule return ErrUsage", theory(
-		when{
-			knitId: "test-Id",
-			flag: lineage.Flag{
-				Upstream:   false,
-				Downstream: false,
-				Numbers:    "abc",
+				Numbers:    pointer.Ref(args.NewDepth(0)),
 			},
 			err: nil,
 		},
@@ -242,7 +224,7 @@ func TestLineageCommand(t *testing.T) {
 			flag: lineage.Flag{
 				Upstream:   false,
 				Downstream: false,
-				Numbers:    "3",
+				Numbers:    pointer.Ref(args.NewDepth(3)),
 			},
 			err: err,
 		},
@@ -255,14 +237,14 @@ func TestLineageCommand(t *testing.T) {
 func TestTraceDownStream(t *testing.T) {
 	type When struct {
 		RootKnitId      string
-		Depth           int
-		ArgGraph        *lineage.DirectedGraph
+		Depth           args.Depth
+		ArgGraph        *knitgraph.DirectedGraph
 		FindDataReturns [][]data.Detail
 		GetRunReturns   []runs.Detail
 	}
 
 	type Then struct {
-		Graph lineage.DirectedGraph
+		Graph knitgraph.DirectedGraph
 		Err   error
 	}
 
@@ -302,7 +284,7 @@ func TestTraceDownStream(t *testing.T) {
 			if !cmp.MapEqWith(
 				graph.DataNodes,
 				then.Graph.DataNodes,
-				func(a, b lineage.DataNode) bool { return a.Equal(&b) },
+				func(a, b knitgraph.DataNode) bool { return a.Equal(&b) },
 			) {
 				t.Errorf(
 					"DataNodes is not equal (actual,expected): %v,%v",
@@ -312,7 +294,7 @@ func TestTraceDownStream(t *testing.T) {
 			if !cmp.MapEqWith(
 				graph.RunNodes,
 				then.Graph.RunNodes,
-				func(a, b lineage.RunNode) bool { return a.Summary.Equal(b.Summary) },
+				func(a, b knitgraph.RunNode) bool { return a.Summary.Equal(b.Summary) },
 			) {
 				t.Errorf(
 					"RunNodes is not equal (actual,expected): %v,%v",
@@ -322,7 +304,7 @@ func TestTraceDownStream(t *testing.T) {
 			if !cmp.MapEqWith(
 				graph.EdgesFromData,
 				then.Graph.EdgesFromData,
-				func(a []lineage.Edge, b []lineage.Edge) bool {
+				func(a []knitgraph.Edge, b []knitgraph.Edge) bool {
 					return cmp.SliceContentEq(a, b)
 				},
 			) {
@@ -334,7 +316,7 @@ func TestTraceDownStream(t *testing.T) {
 			if !cmp.MapEqWith(
 				graph.EdgesFromRun,
 				then.Graph.EdgesFromRun,
-				func(a []lineage.Edge, b []lineage.Edge) bool {
+				func(a []knitgraph.Edge, b []knitgraph.Edge) bool {
 					return cmp.SliceContentEq(a, b)
 				},
 			) {
@@ -364,17 +346,17 @@ func TestTraceDownStream(t *testing.T) {
 		t.Run("Confirm that all the above nodes can be traced in graph depth:1", theory(
 			When{
 				RootKnitId:      "data1",
-				Depth:           1,
-				ArgGraph:        lineage.NewDirectedGraph(),
+				Depth:           args.NewDepth(1),
+				ArgGraph:        knitgraph.NewDirectedGraph(),
 				FindDataReturns: [][]data.Detail{{data1}, {data2}},
 				GetRunReturns:   []runs.Detail{run1},
 			},
 			Then{
-				Graph: lineage.DirectedGraph{
-					DataNodes:     map[string]lineage.DataNode{"data1": toDataNode(data1), "data2": toDataNode(data2)},
-					RunNodes:      map[string]lineage.RunNode{"run1": {run1.Summary}},
-					EdgesFromData: map[string][]lineage.Edge{"data1": {{ToId: "run1", Label: "in/1"}}},
-					EdgesFromRun:  map[string][]lineage.Edge{"run1": {{ToId: "data2", Label: "out/1"}}},
+				Graph: knitgraph.DirectedGraph{
+					DataNodes:     map[string]knitgraph.DataNode{"data1": toDataNode(data1), "data2": toDataNode(data2)},
+					RunNodes:      map[string]knitgraph.RunNode{"run1": {Summary: run1.Summary}},
+					EdgesFromData: map[string][]knitgraph.Edge{"data1": {{ToId: "run1", Label: "in/1"}}},
+					EdgesFromRun:  map[string][]knitgraph.Edge{"run1": {{ToId: "data2", Label: "out/1"}}},
 					RootNodes:     []string{},
 				},
 				Err: nil,
@@ -384,17 +366,17 @@ func TestTraceDownStream(t *testing.T) {
 		t.Run("Confirm that the tracing result remains the same when graph depth is no limit", theory(
 			When{
 				RootKnitId:      "data1",
-				Depth:           -1,
-				ArgGraph:        lineage.NewDirectedGraph(),
+				Depth:           args.NewInfinityDepth(),
+				ArgGraph:        knitgraph.NewDirectedGraph(),
 				FindDataReturns: [][]data.Detail{{data1}, {data2}},
 				GetRunReturns:   []runs.Detail{run1},
 			},
 			Then{
-				Graph: lineage.DirectedGraph{
-					DataNodes:     map[string]lineage.DataNode{"data1": toDataNode(data1), "data2": toDataNode(data2)},
-					RunNodes:      map[string]lineage.RunNode{"run1": {run1.Summary}},
-					EdgesFromData: map[string][]lineage.Edge{"data1": {{ToId: "run1", Label: "in/1"}}},
-					EdgesFromRun:  map[string][]lineage.Edge{"run1": {{ToId: "data2", Label: "out/1"}}},
+				Graph: knitgraph.DirectedGraph{
+					DataNodes:     map[string]knitgraph.DataNode{"data1": toDataNode(data1), "data2": toDataNode(data2)},
+					RunNodes:      map[string]knitgraph.RunNode{"run1": {Summary: run1.Summary}},
+					EdgesFromData: map[string][]knitgraph.Edge{"data1": {{ToId: "run1", Label: "in/1"}}},
+					EdgesFromRun:  map[string][]knitgraph.Edge{"run1": {{ToId: "data2", Label: "out/1"}}},
 					RootNodes:     []string{},
 				},
 				Err: nil,
@@ -414,24 +396,24 @@ func TestTraceDownStream(t *testing.T) {
 		t.Run("Confirm that all nodes can be traced in graph depth:2", theory(
 			When{
 				RootKnitId:      "data1",
-				Depth:           2,
-				ArgGraph:        lineage.NewDirectedGraph(),
+				Depth:           args.NewDepth(2),
+				ArgGraph:        knitgraph.NewDirectedGraph(),
 				FindDataReturns: [][]data.Detail{{data1}, {data2}, {data3}, {data4}},
 				GetRunReturns:   []runs.Detail{run1, run2},
 			},
 			Then{
-				Graph: lineage.DirectedGraph{
-					DataNodes: map[string]lineage.DataNode{
+				Graph: knitgraph.DirectedGraph{
+					DataNodes: map[string]knitgraph.DataNode{
 						"data1": toDataNode(data1), "data2": toDataNode(data2), "data3": toDataNode(data3),
 						"data4": toDataNode(data4),
 					},
-					RunNodes: map[string]lineage.RunNode{
-						"run1": {run1.Summary}, "run2": {run2.Summary},
+					RunNodes: map[string]knitgraph.RunNode{
+						"run1": {Summary: run1.Summary}, "run2": {Summary: run2.Summary},
 					},
-					EdgesFromData: map[string][]lineage.Edge{"data1": {{"run1", "in/1"}}, "data2": {{"run2", "in/2"}}},
-					EdgesFromRun: map[string][]lineage.Edge{
-						"run1": {{"data2", "out/1"}},
-						"run2": {{"data3", "out/2"}, {"data4", "out/3"}},
+					EdgesFromData: map[string][]knitgraph.Edge{"data1": {{ToId: "run1", Label: "in/1"}}, "data2": {{ToId: "run2", Label: "in/2"}}},
+					EdgesFromRun: map[string][]knitgraph.Edge{
+						"run1": {{ToId: "data2", Label: "out/1"}},
+						"run2": {{ToId: "data3", Label: "out/2"}, {ToId: "data4", Label: "out/3"}},
 					},
 					RootNodes: []string{},
 				},
@@ -442,17 +424,17 @@ func TestTraceDownStream(t *testing.T) {
 		t.Run("Confirm that it traces up to data2 in graph depth:1", theory(
 			When{
 				RootKnitId:      "data1",
-				Depth:           1,
-				ArgGraph:        lineage.NewDirectedGraph(),
+				Depth:           args.NewDepth(1),
+				ArgGraph:        knitgraph.NewDirectedGraph(),
 				FindDataReturns: [][]data.Detail{{data1}, {data2}},
 				GetRunReturns:   []runs.Detail{run1},
 			},
 			Then{
-				Graph: lineage.DirectedGraph{
-					DataNodes:     map[string]lineage.DataNode{"data1": toDataNode(data1), "data2": toDataNode(data2)},
-					RunNodes:      map[string]lineage.RunNode{"run1": {run1.Summary}},
-					EdgesFromData: map[string][]lineage.Edge{"data1": {{ToId: "run1", Label: "in/1"}}},
-					EdgesFromRun:  map[string][]lineage.Edge{"run1": {{ToId: "data2", Label: "out/1"}}},
+				Graph: knitgraph.DirectedGraph{
+					DataNodes:     map[string]knitgraph.DataNode{"data1": toDataNode(data1), "data2": toDataNode(data2)},
+					RunNodes:      map[string]knitgraph.RunNode{"run1": {Summary: run1.Summary}},
+					EdgesFromData: map[string][]knitgraph.Edge{"data1": {{ToId: "run1", Label: "in/1"}}},
+					EdgesFromRun:  map[string][]knitgraph.Edge{"run1": {{ToId: "data2", Label: "out/1"}}},
 					RootNodes:     []string{},
 				},
 				Err: nil,
@@ -473,18 +455,18 @@ func TestTraceDownStream(t *testing.T) {
 		t.Run("Confirm that Nodes except run2 can be traced in graph depth: no limit", theory(
 			When{
 				RootKnitId:      "data1",
-				Depth:           -1,
-				ArgGraph:        lineage.NewDirectedGraph(),
+				Depth:           args.NewInfinityDepth(),
+				ArgGraph:        knitgraph.NewDirectedGraph(),
 				FindDataReturns: [][]data.Detail{{data1}, {data2}, {data3}},
 				GetRunReturns:   []runs.Detail{run1}},
 			Then{
-				Graph: lineage.DirectedGraph{
-					DataNodes: map[string]lineage.DataNode{
+				Graph: knitgraph.DirectedGraph{
+					DataNodes: map[string]knitgraph.DataNode{
 						"data1": toDataNode(data1), "data2": toDataNode(data2), "data3": toDataNode(data3),
 					},
-					RunNodes:      map[string]lineage.RunNode{"run1": {run1.Summary}},
-					EdgesFromData: map[string][]lineage.Edge{"data1": {{"run1", "in/1"}}, "data2": {{"run1", "in/2"}}},
-					EdgesFromRun:  map[string][]lineage.Edge{"run1": {{"data3", "out/1"}}},
+					RunNodes:      map[string]knitgraph.RunNode{"run1": {Summary: run1.Summary}},
+					EdgesFromData: map[string][]knitgraph.Edge{"data1": {{ToId: "run1", Label: "in/1"}}, "data2": {{ToId: "run1", Label: "in/2"}}},
+					EdgesFromRun:  map[string][]knitgraph.Edge{"run1": {{ToId: "data3", Label: "out/1"}}},
 					RootNodes:     []string{},
 				},
 				Err: nil,
@@ -506,26 +488,26 @@ func TestTraceDownStream(t *testing.T) {
 		t.Run("Confirm that all nodes can be traced", theory(
 			When{
 				RootKnitId:      "data1",
-				Depth:           -1,
-				ArgGraph:        lineage.NewDirectedGraph(),
+				Depth:           args.NewInfinityDepth(),
+				ArgGraph:        knitgraph.NewDirectedGraph(),
 				FindDataReturns: [][]data.Detail{{data1}, {data2}, {data3}, {data4}},
 				GetRunReturns:   []runs.Detail{run1, run2, run3},
 			},
 			Then{
-				Graph: lineage.DirectedGraph{
-					DataNodes: map[string]lineage.DataNode{
+				Graph: knitgraph.DirectedGraph{
+					DataNodes: map[string]knitgraph.DataNode{
 						"data1": toDataNode(data1), "data2": toDataNode(data2),
 						"data3": toDataNode(data3), "data4": toDataNode(data4),
 					},
-					RunNodes: map[string]lineage.RunNode{
-						"run1": {run1.Summary}, "run2": {run2.Summary}, "run3": {run3.Summary},
+					RunNodes: map[string]knitgraph.RunNode{
+						"run1": {Summary: run1.Summary}, "run2": {Summary: run2.Summary}, "run3": {Summary: run3.Summary},
 					},
-					EdgesFromData: map[string][]lineage.Edge{
-						"data1": {{"run1", "in/1"}, {"run2", "in/2"}},
-						"data2": {{"run3", "in/3"}}, "data3": {{"run3", "in/4"}},
+					EdgesFromData: map[string][]knitgraph.Edge{
+						"data1": {{ToId: "run1", Label: "in/1"}, {ToId: "run2", Label: "in/2"}},
+						"data2": {{ToId: "run3", Label: "in/3"}}, "data3": {{ToId: "run3", Label: "in/4"}},
 					},
-					EdgesFromRun: map[string][]lineage.Edge{
-						"run1": {{"data2", "out/1"}}, "run2": {{"data3", "out/2"}}, "run3": {{"data4", "out/3"}},
+					EdgesFromRun: map[string][]knitgraph.Edge{
+						"run1": {{ToId: "data2", Label: "out/1"}}, "run2": {{ToId: "data3", Label: "out/2"}}, "run3": {{ToId: "data4", Label: "out/3"}},
 					},
 					RootNodes: []string{},
 				},
@@ -549,24 +531,24 @@ func TestTraceDownStream(t *testing.T) {
 		t.Run("Confirm that it traces only downstreams and direct inputs in downstream runs", theory(
 			When{
 				RootKnitId:      "data1",
-				Depth:           -1,
-				ArgGraph:        lineage.NewDirectedGraph(),
+				Depth:           args.NewInfinityDepth(),
+				ArgGraph:        knitgraph.NewDirectedGraph(),
 				FindDataReturns: [][]data.Detail{{data1}, {data2}, {data3}, {data4}},
 				GetRunReturns:   []runs.Detail{run1, run2},
 			},
 			Then{
-				Graph: lineage.DirectedGraph{
-					DataNodes: map[string]lineage.DataNode{
+				Graph: knitgraph.DirectedGraph{
+					DataNodes: map[string]knitgraph.DataNode{
 						"data1": toDataNode(data1), "data2": toDataNode(data2),
 						"data3": toDataNode(data3), "data4": toDataNode(data4),
 					},
-					RunNodes: map[string]lineage.RunNode{"run1": {run1.Summary}, "run2": {run2.Summary}},
-					EdgesFromData: map[string][]lineage.Edge{
-						"data1": {{"run1", "in/1"}}, "data2": {{"run1", "in/2"}, {"run2", "in/3"}},
-						"data3": {{"run2", "in/4"}},
+					RunNodes: map[string]knitgraph.RunNode{"run1": {Summary: run1.Summary}, "run2": {Summary: run2.Summary}},
+					EdgesFromData: map[string][]knitgraph.Edge{
+						"data1": {{ToId: "run1", Label: "in/1"}}, "data2": {{ToId: "run1", Label: "in/2"}, {ToId: "run2", Label: "in/3"}},
+						"data3": {{ToId: "run2", Label: "in/4"}},
 					},
-					EdgesFromRun: map[string][]lineage.Edge{
-						"run1": {{"data3", "out/1"}}, "run2": {{"data4", "out/2"}},
+					EdgesFromRun: map[string][]knitgraph.Edge{
+						"run1": {{ToId: "data3", Label: "out/1"}}, "run2": {{ToId: "data4", Label: "out/2"}},
 					},
 					RootNodes: []string{},
 				},
@@ -594,22 +576,22 @@ func TestTraceDownStream(t *testing.T) {
 		t.Run("Confirm that nodes except run2,run4 and data5 can be traced in graph depth:1", theory(
 			When{
 				RootKnitId:      "data1",
-				Depth:           1,
-				ArgGraph:        lineage.NewDirectedGraph(),
+				Depth:           args.NewDepth(1),
+				ArgGraph:        knitgraph.NewDirectedGraph(),
 				FindDataReturns: [][]data.Detail{{data1}, {data2}, {data3}, {data4}},
 				GetRunReturns:   []runs.Detail{run1, run3},
 			},
 			Then{
-				Graph: lineage.DirectedGraph{
-					DataNodes: map[string]lineage.DataNode{
+				Graph: knitgraph.DirectedGraph{
+					DataNodes: map[string]knitgraph.DataNode{
 						"data1": toDataNode(data1), "data2": toDataNode(data2),
 						"data3": toDataNode(data3), "data4": toDataNode(data4),
 					},
-					RunNodes: map[string]lineage.RunNode{"run1": {run1.Summary}, "run3": {run3.Summary}},
-					EdgesFromData: map[string][]lineage.Edge{
-						"data1": {{"run1", "in/1"}, {"run3", "in/2"}}, "data3": {{"run3", "in/4"}},
+					RunNodes: map[string]knitgraph.RunNode{"run1": {Summary: run1.Summary}, "run3": {Summary: run3.Summary}},
+					EdgesFromData: map[string][]knitgraph.Edge{
+						"data1": {{ToId: "run1", Label: "in/1"}, {ToId: "run3", Label: "in/2"}}, "data3": {{ToId: "run3", Label: "in/4"}},
 					},
-					EdgesFromRun: map[string][]lineage.Edge{"run1": {{"data2", "out/1"}}, "run3": {{"data4", "out/3"}}},
+					EdgesFromRun: map[string][]knitgraph.Edge{"run1": {{ToId: "data2", Label: "out/1"}}, "run3": {{ToId: "data4", Label: "out/3"}}},
 					RootNodes:    []string{},
 				},
 				Err: nil,
@@ -619,26 +601,26 @@ func TestTraceDownStream(t *testing.T) {
 		t.Run("Confirm that nodes except run4 and data5 can be traced in graph depth:2", theory(
 			When{
 				RootKnitId:      "data1",
-				Depth:           2,
-				ArgGraph:        lineage.NewDirectedGraph(),
+				Depth:           args.NewDepth(2),
+				ArgGraph:        knitgraph.NewDirectedGraph(),
 				FindDataReturns: [][]data.Detail{{data1}, {data2}, {data3}, {data4}},
 				GetRunReturns:   []runs.Detail{run1, run3, run2},
 			},
 			Then{
-				Graph: lineage.DirectedGraph{
-					DataNodes: map[string]lineage.DataNode{
+				Graph: knitgraph.DirectedGraph{
+					DataNodes: map[string]knitgraph.DataNode{
 						"data1": toDataNode(data1), "data2": toDataNode(data2),
 						"data3": toDataNode(data3), "data4": toDataNode(data4),
 					},
-					RunNodes: map[string]lineage.RunNode{
-						"run1": {run1.Summary}, "run3": {run3.Summary}, "run2": {run2.Summary},
+					RunNodes: map[string]knitgraph.RunNode{
+						"run1": {Summary: run1.Summary}, "run3": {Summary: run3.Summary}, "run2": {Summary: run2.Summary},
 					},
-					EdgesFromData: map[string][]lineage.Edge{
-						"data1": {{"run1", "in/1"}, {"run3", "in/2"}},
-						"data2": {{"run2", "in/3"}}, "data3": {{"run3", "in/4"}},
+					EdgesFromData: map[string][]knitgraph.Edge{
+						"data1": {{ToId: "run1", Label: "in/1"}, {ToId: "run3", Label: "in/2"}},
+						"data2": {{ToId: "run2", Label: "in/3"}}, "data3": {{ToId: "run3", Label: "in/4"}},
 					},
-					EdgesFromRun: map[string][]lineage.Edge{
-						"run1": {{"data2", "out/1"}}, "run2": {{"data3", "out/2"}}, "run3": {{"data4", "out/3"}},
+					EdgesFromRun: map[string][]knitgraph.Edge{
+						"run1": {{ToId: "data2", Label: "out/1"}}, "run2": {{ToId: "data3", Label: "out/2"}}, "run3": {{ToId: "data4", Label: "out/3"}},
 					},
 					RootNodes: []string{},
 				},
@@ -649,29 +631,29 @@ func TestTraceDownStream(t *testing.T) {
 		t.Run("Confirm that all nodes can be traced in graph depth:3", theory(
 			When{
 				RootKnitId:      "data1",
-				Depth:           3,
-				ArgGraph:        lineage.NewDirectedGraph(),
+				Depth:           args.NewDepth(3),
+				ArgGraph:        knitgraph.NewDirectedGraph(),
 				FindDataReturns: [][]data.Detail{{data1}, {data2}, {data3}, {data4}, {data5}},
 				GetRunReturns:   []runs.Detail{run1, run3, run2, run4},
 			},
 			Then{
-				Graph: lineage.DirectedGraph{
-					DataNodes: map[string]lineage.DataNode{
+				Graph: knitgraph.DirectedGraph{
+					DataNodes: map[string]knitgraph.DataNode{
 						"data1": toDataNode(data1), "data2": toDataNode(data2), "data3": toDataNode(data3),
 						"data4": toDataNode(data4), "data5": toDataNode(data5),
 					},
-					RunNodes: map[string]lineage.RunNode{
-						"run1": {run1.Summary}, "run3": {run3.Summary},
-						"run2": {run2.Summary}, "run4": {run4.Summary},
+					RunNodes: map[string]knitgraph.RunNode{
+						"run1": {Summary: run1.Summary}, "run3": {Summary: run3.Summary},
+						"run2": {Summary: run2.Summary}, "run4": {Summary: run4.Summary},
 					},
-					EdgesFromData: map[string][]lineage.Edge{
-						"data1": {{"run1", "in/1"}, {"run3", "in/2"}},
-						"data2": {{"run2", "in/3"}},
-						"data3": {{"run3", "in/4"}, {"run4", "in/5"}},
+					EdgesFromData: map[string][]knitgraph.Edge{
+						"data1": {{ToId: "run1", Label: "in/1"}, {ToId: "run3", Label: "in/2"}},
+						"data2": {{ToId: "run2", Label: "in/3"}},
+						"data3": {{ToId: "run3", Label: "in/4"}, {ToId: "run4", Label: "in/5"}},
 					},
-					EdgesFromRun: map[string][]lineage.Edge{
-						"run1": {{"data2", "out/1"}}, "run2": {{"data3", "out/2"}},
-						"run3": {{"data4", "out/3"}}, "run4": {{"data5", "out/4"}},
+					EdgesFromRun: map[string][]knitgraph.Edge{
+						"run1": {{ToId: "data2", Label: "out/1"}}, "run2": {{ToId: "data3", Label: "out/2"}},
+						"run3": {{ToId: "data4", Label: "out/3"}}, "run4": {{ToId: "data5", Label: "out/4"}},
 					},
 					RootNodes: []string{},
 				},
@@ -689,17 +671,17 @@ func TestTraceDownStream(t *testing.T) {
 		t.Run("confirm that all nodes can be traced even when the run does not have an output", theory(
 			When{
 				RootKnitId:      "data1",
-				Depth:           -1,
-				ArgGraph:        lineage.NewDirectedGraph(),
+				Depth:           args.NewInfinityDepth(),
+				ArgGraph:        knitgraph.NewDirectedGraph(),
 				FindDataReturns: [][]data.Detail{{data1}, {data2}},
 				GetRunReturns:   []runs.Detail{run1, run2},
 			},
 			Then{
-				Graph: lineage.DirectedGraph{
-					DataNodes:     map[string]lineage.DataNode{"data1": toDataNode(data1), "data2": toDataNode(data2)},
-					RunNodes:      map[string]lineage.RunNode{"run1": {run1.Summary}, "run2": {run2.Summary}},
-					EdgesFromData: map[string][]lineage.Edge{"data1": {{"run1", "in/1"}}, "data2": {{"run2", "in/2"}}},
-					EdgesFromRun:  map[string][]lineage.Edge{"run1": {{"data2", "out/1"}}},
+				Graph: knitgraph.DirectedGraph{
+					DataNodes:     map[string]knitgraph.DataNode{"data1": toDataNode(data1), "data2": toDataNode(data2)},
+					RunNodes:      map[string]knitgraph.RunNode{"run1": {Summary: run1.Summary}, "run2": {Summary: run2.Summary}},
+					EdgesFromData: map[string][]knitgraph.Edge{"data1": {{ToId: "run1", Label: "in/1"}}, "data2": {{ToId: "run2", Label: "in/2"}}},
+					EdgesFromRun:  map[string][]knitgraph.Edge{"run1": {{ToId: "data2", Label: "out/1"}}},
 					RootNodes:     []string{},
 				},
 				Err: nil,
@@ -724,15 +706,15 @@ func TestTraceDownStream(t *testing.T) {
 		t.Run("Confirm that all nodes can be obtained by tracing both upstream and downstream.", theory(
 			When{
 				RootKnitId: "data1",
-				Depth:      -1,
-				ArgGraph: &lineage.DirectedGraph{
-					DataNodes: map[string]lineage.DataNode{"data1": toDataNode(data1), "data2": toDataNode(data2), "data3": toDataNode(data3)},
-					RunNodes:  map[string]lineage.RunNode{"run1": {run1.Summary}},
-					EdgesFromData: map[string][]lineage.Edge{
-						"data3": {{"run1", "in/1"}},
+				Depth:      args.NewInfinityDepth(),
+				ArgGraph: &knitgraph.DirectedGraph{
+					DataNodes: map[string]knitgraph.DataNode{"data1": toDataNode(data1), "data2": toDataNode(data2), "data3": toDataNode(data3)},
+					RunNodes:  map[string]knitgraph.RunNode{"run1": {Summary: run1.Summary}},
+					EdgesFromData: map[string][]knitgraph.Edge{
+						"data3": {{ToId: "run1", Label: "in/1"}},
 					},
-					EdgesFromRun: map[string][]lineage.Edge{
-						"run1": {{"data1", "out/1"}, {"data2", "out/2"}},
+					EdgesFromRun: map[string][]knitgraph.Edge{
+						"run1": {{ToId: "data1", Label: "out/1"}, {ToId: "data2", Label: "out/2"}},
 					},
 					RootNodes: []string{},
 				},
@@ -740,19 +722,19 @@ func TestTraceDownStream(t *testing.T) {
 				GetRunReturns:   []runs.Detail{run2},
 			},
 			Then{
-				Graph: lineage.DirectedGraph{
-					DataNodes: map[string]lineage.DataNode{
+				Graph: knitgraph.DirectedGraph{
+					DataNodes: map[string]knitgraph.DataNode{
 						"data1": toDataNode(data1), "data2": toDataNode(data2),
 						"data3": toDataNode(data3), "data4": toDataNode(data4),
 					},
-					RunNodes: map[string]lineage.RunNode{"run1": {run1.Summary}, "run2": {run2.Summary}},
-					EdgesFromData: map[string][]lineage.Edge{
-						"data1": {{"run2", "in/2"}}, "data2": {{"run2", "in/3"}},
-						"data3": {{"run1", "in/1"}},
+					RunNodes: map[string]knitgraph.RunNode{"run1": {Summary: run1.Summary}, "run2": {Summary: run2.Summary}},
+					EdgesFromData: map[string][]knitgraph.Edge{
+						"data1": {{ToId: "run2", Label: "in/2"}}, "data2": {{ToId: "run2", Label: "in/3"}},
+						"data3": {{ToId: "run1", Label: "in/1"}},
 					},
-					EdgesFromRun: map[string][]lineage.Edge{
-						"run1": {{"data1", "out/1"}, {"data2", "out/2"}},
-						"run2": {{"data4", "out/3"}},
+					EdgesFromRun: map[string][]knitgraph.Edge{
+						"run1": {{ToId: "data1", Label: "out/1"}, {ToId: "data2", Label: "out/2"}},
+						"run2": {{ToId: "data4", Label: "out/3"}},
 					},
 					RootNodes: []string{},
 				},
@@ -777,26 +759,26 @@ func TestTraceDownStream(t *testing.T) {
 		t.Run("Confirm that all nodes containg log can be obtained", theory(
 			When{
 				RootKnitId: "data1",
-				Depth:      -1,
-				ArgGraph:   lineage.NewDirectedGraph(),
+				Depth:      args.NewInfinityDepth(),
+				ArgGraph:   knitgraph.NewDirectedGraph(),
 				FindDataReturns: [][]data.Detail{
 					{data1}, {data2}, {data3}, {data4},
 				},
 				GetRunReturns: []runs.Detail{run1, run2},
 			},
 			Then{
-				Graph: lineage.DirectedGraph{
-					DataNodes: map[string]lineage.DataNode{
+				Graph: knitgraph.DirectedGraph{
+					DataNodes: map[string]knitgraph.DataNode{
 						"data1": toDataNode(data1), "data2": toDataNode(data2),
 						"data3": toDataNode(data3), "data4": toDataNode(data4),
 					},
-					RunNodes: map[string]lineage.RunNode{"run1": {run1.Summary}, "run2": {run2.Summary}},
-					EdgesFromData: map[string][]lineage.Edge{
-						"data1": {{"run1", "in/1"}}, "data3": {{"run2", "in/2"}},
+					RunNodes: map[string]knitgraph.RunNode{"run1": {Summary: run1.Summary}, "run2": {Summary: run2.Summary}},
+					EdgesFromData: map[string][]knitgraph.Edge{
+						"data1": {{ToId: "run1", Label: "in/1"}}, "data3": {{ToId: "run2", Label: "in/2"}},
 					},
-					EdgesFromRun: map[string][]lineage.Edge{
-						"run1": {{"data2", "out/1"}, {"data3", "(log)"}},
-						"run2": {{"data4", "out/2"}},
+					EdgesFromRun: map[string][]knitgraph.Edge{
+						"run1": {{ToId: "data2", Label: "out/1"}, {ToId: "data3", Label: "(log)"}},
+						"run2": {{ToId: "data4", Label: "out/2"}},
 					},
 					RootNodes: []string{},
 				},
@@ -815,9 +797,9 @@ func TestTraceDownStream(t *testing.T) {
 		}
 
 		ctx := context.Background()
-		graph := lineage.NewDirectedGraph()
+		graph := knitgraph.NewDirectedGraph()
 		knitId := "knitId-test"
-		_, actual := lineage.TraceDownStream(ctx, mock, graph, knitId, 1)
+		_, actual := lineage.TraceDownStream(ctx, mock, graph, knitId, args.NewDepth(1))
 		if !errors.Is(actual, lineage.ErrNotFoundData) {
 			t.Errorf("wrong status: (actual, expected) != (%v, %v)", actual, lineage.ErrNotFoundData)
 		}
@@ -834,9 +816,9 @@ func TestTraceDownStream(t *testing.T) {
 		}
 
 		ctx := context.Background()
-		graph := lineage.NewDirectedGraph()
+		graph := knitgraph.NewDirectedGraph()
 		knitId := "knitId-test"
-		_, actual := lineage.TraceDownStream(ctx, mock, graph, knitId, 1)
+		_, actual := lineage.TraceDownStream(ctx, mock, graph, knitId, args.NewDepth(1))
 		if !errors.Is(actual, expectedError) {
 			t.Errorf("wrong status: (actual, expected) != (%v, %v)", actual, expectedError)
 		}
@@ -854,8 +836,8 @@ func TestTraceDownStream(t *testing.T) {
 			return runs.Detail{}, expectedError
 		}
 		ctx := context.Background()
-		graph := lineage.NewDirectedGraph()
-		_, actual := lineage.TraceDownStream(ctx, mock, graph, knitId, 1)
+		graph := knitgraph.NewDirectedGraph()
+		_, actual := lineage.TraceDownStream(ctx, mock, graph, knitId, args.NewDepth(1))
 		if !errors.Is(actual, expectedError) {
 			t.Errorf("wrong status: (actual, expected) != (%s, %d)", actual, expectedError)
 		}
@@ -865,13 +847,13 @@ func TestTraceDownStream(t *testing.T) {
 func TestTraceUpStream(t *testing.T) {
 	type When struct {
 		RootKnitId      string
-		Depth           int
+		Depth           args.Depth
 		FindDataReturns [][]data.Detail
 		GetRunReturns   []runs.Detail
 	}
 
 	type Then struct {
-		Graph lineage.DirectedGraph
+		Graph knitgraph.DirectedGraph
 		Err   error
 	}
 
@@ -900,7 +882,7 @@ func TestTraceUpStream(t *testing.T) {
 			}
 
 			ctx := context.Background()
-			graph := lineage.NewDirectedGraph()
+			graph := knitgraph.NewDirectedGraph()
 			graph, actual := lineage.TraceUpStream(ctx, mock, graph, when.RootKnitId, when.Depth)
 			if !errors.Is(actual, then.Err) {
 				t.Errorf(
@@ -912,7 +894,7 @@ func TestTraceUpStream(t *testing.T) {
 			if !cmp.MapEqWith(
 				graph.DataNodes,
 				then.Graph.DataNodes,
-				func(a, b lineage.DataNode) bool { return a.Equal(&b) },
+				func(a, b knitgraph.DataNode) bool { return a.Equal(&b) },
 			) {
 				t.Errorf(
 					"DataNodes is not equal (actual,expected): %v,%v",
@@ -922,7 +904,7 @@ func TestTraceUpStream(t *testing.T) {
 			if !cmp.MapEqWith(
 				graph.RunNodes,
 				then.Graph.RunNodes,
-				func(a, b lineage.RunNode) bool { return a.Summary.Equal(b.Summary) },
+				func(a, b knitgraph.RunNode) bool { return a.Summary.Equal(b.Summary) },
 			) {
 				t.Errorf(
 					"RunNodes is not equal (actual,expected): %v,%v",
@@ -932,7 +914,7 @@ func TestTraceUpStream(t *testing.T) {
 			if !cmp.MapEqWith(
 				graph.EdgesFromData,
 				then.Graph.EdgesFromData,
-				func(a []lineage.Edge, b []lineage.Edge) bool {
+				func(a []knitgraph.Edge, b []knitgraph.Edge) bool {
 					return cmp.SliceContentEq(a, b)
 				},
 			) {
@@ -944,7 +926,7 @@ func TestTraceUpStream(t *testing.T) {
 			if !cmp.MapEqWith(
 				graph.EdgesFromRun,
 				then.Graph.EdgesFromRun,
-				func(a []lineage.Edge, b []lineage.Edge) bool {
+				func(a []knitgraph.Edge, b []knitgraph.Edge) bool {
 					return cmp.SliceContentEq(a, b)
 				},
 			) {
@@ -972,16 +954,16 @@ func TestTraceUpStream(t *testing.T) {
 		t.Run("Confirm that all the above nodes can be traced in graph depth:1", theory(
 			When{
 				RootKnitId:      "data1",
-				Depth:           1,
+				Depth:           args.NewDepth(1),
 				FindDataReturns: [][]data.Detail{{data1}},
 				GetRunReturns:   []runs.Detail{run1},
 			},
 			Then{
-				Graph: lineage.DirectedGraph{
-					DataNodes:     map[string]lineage.DataNode{"data1": toDataNode(data1)},
-					RunNodes:      map[string]lineage.RunNode{"run1": {run1.Summary}},
-					EdgesFromData: map[string][]lineage.Edge{},
-					EdgesFromRun:  map[string][]lineage.Edge{"run1": {{ToId: "data1", Label: "upload"}}},
+				Graph: knitgraph.DirectedGraph{
+					DataNodes:     map[string]knitgraph.DataNode{"data1": toDataNode(data1)},
+					RunNodes:      map[string]knitgraph.RunNode{"run1": {Summary: run1.Summary}},
+					EdgesFromData: map[string][]knitgraph.Edge{},
+					EdgesFromRun:  map[string][]knitgraph.Edge{"run1": {{ToId: "data1", Label: "upload"}}},
 					RootNodes:     []string{"run1"},
 				},
 				Err: nil,
@@ -990,16 +972,16 @@ func TestTraceUpStream(t *testing.T) {
 		t.Run("Confirm that all the above nodes can be traced in graph depth: no limit", theory(
 			When{
 				RootKnitId:      "data1",
-				Depth:           -1,
+				Depth:           args.NewInfinityDepth(),
 				FindDataReturns: [][]data.Detail{{data1}},
 				GetRunReturns:   []runs.Detail{run1},
 			},
 			Then{
-				Graph: lineage.DirectedGraph{
-					DataNodes:     map[string]lineage.DataNode{"data1": toDataNode(data1)},
-					RunNodes:      map[string]lineage.RunNode{"run1": {run1.Summary}},
-					EdgesFromData: map[string][]lineage.Edge{},
-					EdgesFromRun:  map[string][]lineage.Edge{"run1": {{ToId: "data1", Label: "upload"}}},
+				Graph: knitgraph.DirectedGraph{
+					DataNodes:     map[string]knitgraph.DataNode{"data1": toDataNode(data1)},
+					RunNodes:      map[string]knitgraph.RunNode{"run1": {Summary: run1.Summary}},
+					EdgesFromData: map[string][]knitgraph.Edge{},
+					EdgesFromRun:  map[string][]knitgraph.Edge{"run1": {{ToId: "data1", Label: "upload"}}},
 					RootNodes:     []string{"run1"},
 				},
 				Err: nil,
@@ -1016,20 +998,20 @@ func TestTraceUpStream(t *testing.T) {
 		t.Run("Confirm that all nodes can be traced in graph depth:2", theory(
 			When{
 				RootKnitId:      "data2",
-				Depth:           2,
+				Depth:           args.NewDepth(2),
 				FindDataReturns: [][]data.Detail{{data2}, {data1}},
 				GetRunReturns:   []runs.Detail{run2, run1},
 			},
 			Then{
-				Graph: lineage.DirectedGraph{
-					DataNodes: map[string]lineage.DataNode{
+				Graph: knitgraph.DirectedGraph{
+					DataNodes: map[string]knitgraph.DataNode{
 						"data1": toDataNode(data1), "data2": toDataNode(data2),
 					},
-					RunNodes: map[string]lineage.RunNode{
-						"run1": {run1.Summary}, "run2": {run2.Summary},
+					RunNodes: map[string]knitgraph.RunNode{
+						"run1": {Summary: run1.Summary}, "run2": {Summary: run2.Summary},
 					},
-					EdgesFromData: map[string][]lineage.Edge{"data1": {{"run2", "in/1"}}},
-					EdgesFromRun:  map[string][]lineage.Edge{"run1": {{"data1", "upload"}}, "run2": {{"data2", "out/1"}}},
+					EdgesFromData: map[string][]knitgraph.Edge{"data1": {{ToId: "run2", Label: "in/1"}}},
+					EdgesFromRun:  map[string][]knitgraph.Edge{"run1": {{ToId: "data1", Label: "upload"}}, "run2": {{ToId: "data2", Label: "out/1"}}},
 					RootNodes:     []string{"run1"},
 				},
 				Err: nil,
@@ -1038,20 +1020,20 @@ func TestTraceUpStream(t *testing.T) {
 		t.Run("Confirm that all nodes can be traced up to data1 in graph depth:1", theory(
 			When{
 				RootKnitId:      "data2",
-				Depth:           1,
+				Depth:           args.NewDepth(1),
 				FindDataReturns: [][]data.Detail{{data2}, {data1}},
 				GetRunReturns:   []runs.Detail{run2},
 			},
 			Then{
-				Graph: lineage.DirectedGraph{
-					DataNodes: map[string]lineage.DataNode{
+				Graph: knitgraph.DirectedGraph{
+					DataNodes: map[string]knitgraph.DataNode{
 						"data1": toDataNode(data1), "data2": toDataNode(data2),
 					},
-					RunNodes: map[string]lineage.RunNode{
-						"run2": {run2.Summary},
+					RunNodes: map[string]knitgraph.RunNode{
+						"run2": {Summary: run2.Summary},
 					},
-					EdgesFromData: map[string][]lineage.Edge{"data1": {{"run2", "in/1"}}},
-					EdgesFromRun:  map[string][]lineage.Edge{"run2": {{"data2", "out/1"}}},
+					EdgesFromData: map[string][]knitgraph.Edge{"data1": {{ToId: "run2", Label: "in/1"}}},
+					EdgesFromRun:  map[string][]knitgraph.Edge{"run2": {{ToId: "data2", Label: "out/1"}}},
 					RootNodes:     []string{},
 				},
 				Err: nil,
@@ -1071,21 +1053,21 @@ func TestTraceUpStream(t *testing.T) {
 		t.Run("Confirm it traces only upstreams and direct outputs in upstream runs", theory(
 			When{
 				RootKnitId:      "data3",
-				Depth:           2,
+				Depth:           args.NewDepth(2),
 				FindDataReturns: [][]data.Detail{{data3}, {data2}, {data1}},
 				GetRunReturns:   []runs.Detail{run2, run1},
 			},
 			Then{
-				Graph: lineage.DirectedGraph{
-					DataNodes: map[string]lineage.DataNode{
+				Graph: knitgraph.DirectedGraph{
+					DataNodes: map[string]knitgraph.DataNode{
 						"data1": toDataNode(data1), "data2": toDataNode(data2), "data3": toDataNode(data3),
 					},
-					RunNodes: map[string]lineage.RunNode{
-						"run1": {run1.Summary}, "run2": {run2.Summary},
+					RunNodes: map[string]knitgraph.RunNode{
+						"run1": {Summary: run1.Summary}, "run2": {Summary: run2.Summary},
 					},
-					EdgesFromData: map[string][]lineage.Edge{"data1": {{"run2", "in/1"}}},
-					EdgesFromRun: map[string][]lineage.Edge{
-						"run1": {{"data1", "upload"}}, "run2": {{"data2", "out/1"}, {"data3", "out/2"}},
+					EdgesFromData: map[string][]knitgraph.Edge{"data1": {{ToId: "run2", Label: "in/1"}}},
+					EdgesFromRun: map[string][]knitgraph.Edge{
+						"run1": {{ToId: "data1", Label: "upload"}}, "run2": {{ToId: "data2", Label: "out/1"}, {ToId: "data3", Label: "out/2"}},
 					},
 					RootNodes: []string{"run1"},
 				},
@@ -1108,25 +1090,25 @@ func TestTraceUpStream(t *testing.T) {
 		t.Run("Confirm that all nodes can be traced", theory(
 			When{
 				RootKnitId:      "data4",
-				Depth:           -1,
+				Depth:           args.NewInfinityDepth(),
 				FindDataReturns: [][]data.Detail{{data4}, {data2}, {data3}, {data1}},
 				GetRunReturns:   []runs.Detail{run3, run2, run1},
 			},
 			Then{
-				Graph: lineage.DirectedGraph{
-					DataNodes: map[string]lineage.DataNode{
+				Graph: knitgraph.DirectedGraph{
+					DataNodes: map[string]knitgraph.DataNode{
 						"data1": toDataNode(data1), "data2": toDataNode(data2), "data3": toDataNode(data3),
 						"data4": toDataNode(data4),
 					},
-					RunNodes: map[string]lineage.RunNode{
-						"run1": {run1.Summary}, "run2": {run2.Summary}, "run3": {run3.Summary},
+					RunNodes: map[string]knitgraph.RunNode{
+						"run1": {Summary: run1.Summary}, "run2": {Summary: run2.Summary}, "run3": {Summary: run3.Summary},
 					},
-					EdgesFromData: map[string][]lineage.Edge{
-						"data1": {{"run2", "in/1"}}, "data2": {{"run3", "in/2"}}, "data3": {{"run3", "in/3"}},
+					EdgesFromData: map[string][]knitgraph.Edge{
+						"data1": {{ToId: "run2", Label: "in/1"}}, "data2": {{ToId: "run3", Label: "in/2"}}, "data3": {{ToId: "run3", Label: "in/3"}},
 					},
-					EdgesFromRun: map[string][]lineage.Edge{
-						"run1": {{"data1", "upload"}}, "run2": {{"data2", "out/1"}, {"data3", "out/2"}},
-						"run3": {{"data4", "out/3"}},
+					EdgesFromRun: map[string][]knitgraph.Edge{
+						"run1": {{ToId: "data1", Label: "upload"}}, "run2": {{ToId: "data2", Label: "out/1"}, {ToId: "data3", Label: "out/2"}},
+						"run3": {{ToId: "data4", Label: "out/3"}},
 					},
 					RootNodes: []string{"run1"},
 				},
@@ -1153,23 +1135,23 @@ func TestTraceUpStream(t *testing.T) {
 		t.Run("Confirm that run4, data3 and data1 can be traced in graph depth;1 ", theory(
 			When{
 				RootKnitId:      "data4",
-				Depth:           1,
+				Depth:           args.NewDepth(1),
 				FindDataReturns: [][]data.Detail{{data4}, {data1}, {data3}},
 				GetRunReturns:   []runs.Detail{run4},
 			},
 			Then{
-				Graph: lineage.DirectedGraph{
-					DataNodes: map[string]lineage.DataNode{
+				Graph: knitgraph.DirectedGraph{
+					DataNodes: map[string]knitgraph.DataNode{
 						"data1": toDataNode(data1), "data3": toDataNode(data3), "data4": toDataNode(data4),
 					},
-					RunNodes: map[string]lineage.RunNode{
-						"run4": {run4.Summary},
+					RunNodes: map[string]knitgraph.RunNode{
+						"run4": {Summary: run4.Summary},
 					},
-					EdgesFromData: map[string][]lineage.Edge{
-						"data1": {{"run4", "in/3"}}, "data3": {{"run4", "in/4"}},
+					EdgesFromData: map[string][]knitgraph.Edge{
+						"data1": {{ToId: "run4", Label: "in/3"}}, "data3": {{ToId: "run4", Label: "in/4"}},
 					},
-					EdgesFromRun: map[string][]lineage.Edge{
-						"run4": {{"data4", "out/3"}},
+					EdgesFromRun: map[string][]knitgraph.Edge{
+						"run4": {{ToId: "data4", Label: "out/3"}},
 					},
 					RootNodes: []string{},
 				},
@@ -1179,25 +1161,25 @@ func TestTraceUpStream(t *testing.T) {
 		t.Run("Confirm that run3, data2, run1 and root can be traced in graph depth;2 ", theory(
 			When{
 				RootKnitId:      "data4",
-				Depth:           2,
+				Depth:           args.NewDepth(2),
 				FindDataReturns: [][]data.Detail{{data4}, {data1}, {data3}, {data2}},
 				GetRunReturns:   []runs.Detail{run4, run1, run3},
 			},
 			Then{
-				Graph: lineage.DirectedGraph{
-					DataNodes: map[string]lineage.DataNode{
+				Graph: knitgraph.DirectedGraph{
+					DataNodes: map[string]knitgraph.DataNode{
 						"data1": toDataNode(data1), "data2": toDataNode(data2),
 						"data3": toDataNode(data3), "data4": toDataNode(data4),
 					},
-					RunNodes: map[string]lineage.RunNode{
-						"run1": {run1.Summary}, "run3": {run3.Summary}, "run4": {run4.Summary},
+					RunNodes: map[string]knitgraph.RunNode{
+						"run1": {Summary: run1.Summary}, "run3": {Summary: run3.Summary}, "run4": {Summary: run4.Summary},
 					},
-					EdgesFromData: map[string][]lineage.Edge{
-						"data1": {{"run4", "in/3"}, {"run3", "in/2"}},
-						"data2": {{"run3", "in/4"}}, "data3": {{"run4", "in/4"}},
+					EdgesFromData: map[string][]knitgraph.Edge{
+						"data1": {{ToId: "run4", Label: "in/3"}, {ToId: "run3", Label: "in/2"}},
+						"data2": {{ToId: "run3", Label: "in/4"}}, "data3": {{ToId: "run4", Label: "in/4"}},
 					},
-					EdgesFromRun: map[string][]lineage.Edge{
-						"run4": {{"data4", "out/3"}}, "run3": {{"data3", "out/2"}}, "run1": {{"data1", "upload"}},
+					EdgesFromRun: map[string][]knitgraph.Edge{
+						"run4": {{ToId: "data4", Label: "out/3"}}, "run3": {{ToId: "data3", Label: "out/2"}}, "run1": {{ToId: "data1", Label: "upload"}},
 					},
 					RootNodes: []string{"run1"},
 				},
@@ -1208,27 +1190,27 @@ func TestTraceUpStream(t *testing.T) {
 		t.Run("Confirm that all nodes root can be traced in graph depth;3 ", theory(
 			When{
 				RootKnitId:      "data4",
-				Depth:           3,
+				Depth:           args.NewDepth(3),
 				FindDataReturns: [][]data.Detail{{data4}, {data1}, {data3}, {data2}},
 				GetRunReturns:   []runs.Detail{run4, run1, run3, run2},
 			},
 			Then{
-				Graph: lineage.DirectedGraph{
-					DataNodes: map[string]lineage.DataNode{
+				Graph: knitgraph.DirectedGraph{
+					DataNodes: map[string]knitgraph.DataNode{
 						"data1": toDataNode(data1), "data2": toDataNode(data2),
 						"data3": toDataNode(data3), "data4": toDataNode(data4),
 					},
-					RunNodes: map[string]lineage.RunNode{
-						"run1": {run1.Summary}, "run2": {run2.Summary},
-						"run3": {run3.Summary}, "run4": {run4.Summary},
+					RunNodes: map[string]knitgraph.RunNode{
+						"run1": {Summary: run1.Summary}, "run2": {Summary: run2.Summary},
+						"run3": {Summary: run3.Summary}, "run4": {Summary: run4.Summary},
 					},
-					EdgesFromData: map[string][]lineage.Edge{
-						"data1": {{"run4", "in/3"}, {"run3", "in/2"}, {"run2", "in/1"}},
-						"data2": {{"run3", "in/4"}}, "data3": {{"run4", "in/4"}},
+					EdgesFromData: map[string][]knitgraph.Edge{
+						"data1": {{ToId: "run4", Label: "in/3"}, {ToId: "run3", Label: "in/2"}, {ToId: "run2", Label: "in/1"}},
+						"data2": {{ToId: "run3", Label: "in/4"}}, "data3": {{ToId: "run4", Label: "in/4"}},
 					},
-					EdgesFromRun: map[string][]lineage.Edge{
-						"run4": {{"data4", "out/3"}}, "run3": {{"data3", "out/2"}}, "run1": {{"data1", "upload"}},
-						"run2": {{"data2", "out/1"}},
+					EdgesFromRun: map[string][]knitgraph.Edge{
+						"run4": {{ToId: "data4", Label: "out/3"}}, "run3": {{ToId: "data3", Label: "out/2"}}, "run1": {{ToId: "data1", Label: "upload"}},
+						"run2": {{ToId: "data2", Label: "out/1"}},
 					},
 					RootNodes: []string{"run1"},
 				},
@@ -1256,25 +1238,25 @@ func TestTraceUpStream(t *testing.T) {
 		t.Run("Confirm that all nodes containing log can be obtained.", theory(
 			When{
 				RootKnitId:      "data5",
-				Depth:           2,
+				Depth:           args.NewDepth(2),
 				FindDataReturns: [][]data.Detail{{data5}, {data4}, {data3}, {data2}, {data1}},
 				GetRunReturns:   []runs.Detail{run2, run1},
 			},
 			Then{
-				Graph: lineage.DirectedGraph{
-					DataNodes: map[string]lineage.DataNode{
+				Graph: knitgraph.DirectedGraph{
+					DataNodes: map[string]knitgraph.DataNode{
 						"data1": toDataNode(data1), "data2": toDataNode(data2),
 						"data3": toDataNode(data3), "data4": toDataNode(data4), "data5": toDataNode(data5),
 					},
-					RunNodes: map[string]lineage.RunNode{
-						"run1": {run1.Summary}, "run2": {run2.Summary},
+					RunNodes: map[string]knitgraph.RunNode{
+						"run1": {Summary: run1.Summary}, "run2": {Summary: run2.Summary},
 					},
-					EdgesFromData: map[string][]lineage.Edge{
-						"data1": {{"run1", "in/1"}}, "data3": {{"run2", "in/2"}},
+					EdgesFromData: map[string][]knitgraph.Edge{
+						"data1": {{ToId: "run1", Label: "in/1"}}, "data3": {{ToId: "run2", Label: "in/2"}},
 					},
-					EdgesFromRun: map[string][]lineage.Edge{
-						"run1": {{"data3", "out/1"}, {"data2", "(log)"}},
-						"run2": {{"data4", "out/2"}, {"data5", "(log)"}},
+					EdgesFromRun: map[string][]knitgraph.Edge{
+						"run1": {{ToId: "data3", Label: "out/1"}, {ToId: "data2", Label: "(log)"}},
+						"run2": {{ToId: "data4", Label: "out/2"}, {ToId: "data5", Label: "(log)"}},
 					},
 					RootNodes: []string{},
 				},
@@ -1293,9 +1275,9 @@ func TestTraceUpStream(t *testing.T) {
 		}
 
 		ctx := context.Background()
-		graph := lineage.NewDirectedGraph()
+		graph := knitgraph.NewDirectedGraph()
 		knitId := "knitId-test"
-		_, actual := lineage.TraceUpStream(ctx, mock, graph, knitId, 1)
+		_, actual := lineage.TraceUpStream(ctx, mock, graph, knitId, args.NewDepth(1))
 		if !errors.Is(actual, lineage.ErrNotFoundData) {
 			t.Errorf("wrong status: (actual, expected) != (%v, %v)", actual, lineage.ErrNotFoundData)
 		}
@@ -1312,9 +1294,9 @@ func TestTraceUpStream(t *testing.T) {
 		}
 
 		ctx := context.Background()
-		graph := lineage.NewDirectedGraph()
+		graph := knitgraph.NewDirectedGraph()
 		knitId := "knitId-test"
-		_, actual := lineage.TraceUpStream(ctx, mock, graph, knitId, 1)
+		_, actual := lineage.TraceUpStream(ctx, mock, graph, knitId, args.NewDepth(1))
 		if !errors.Is(actual, expectedError) {
 			t.Errorf("wrong status: (actual, expected) != (%v, %v)", actual, expectedError)
 		}
@@ -1332,16 +1314,16 @@ func TestTraceUpStream(t *testing.T) {
 			}, nil
 		}
 		ctx := context.Background()
-		graph := lineage.NewDirectedGraph()
-		graph, actual := lineage.TraceUpStream(ctx, mock, graph, knitId, 1)
+		graph := knitgraph.NewDirectedGraph()
+		graph, actual := lineage.TraceUpStream(ctx, mock, graph, knitId, args.NewDepth(1))
 		if !errors.Is(actual, expectedError) {
 			t.Errorf("wrong status: (actual, expected) != (%s, %d)", graph.DataNodes, expectedError)
 		}
 	})
 }
 
-func toDataNode(data data.Detail) lineage.DataNode {
-	return lineage.DataNode{
+func toDataNode(data data.Detail) knitgraph.DataNode {
+	return knitgraph.DataNode{
 		KnitId:    data.KnitId,
 		Tags:      data.Tags,
 		FromRunId: data.Upstream.Run.RunId,
@@ -1552,467 +1534,4 @@ func dummySliceAssignment(knitIdToMoutPath map[string]string) []runs.Assignment 
 
 	}
 	return slice
-}
-func TestGenerateDot(t *testing.T) {
-	type When struct {
-		Graph     lineage.DirectedGraph
-		ArgKnitId string
-	}
-	type Then struct {
-		RequiredContent string
-	}
-	theory := func(when When, then Then) func(*testing.T) {
-		return func(t *testing.T) {
-			w := new(strings.Builder)
-
-			if err := when.Graph.GenerateDot(w, when.ArgKnitId); err != nil {
-				t.Fatal(err)
-			}
-
-			if w.String() != then.RequiredContent {
-				t.Errorf("fail \nactual:\n%s \n=========\nexpect:\n%s", w.String(), then.RequiredContent)
-			}
-		}
-	}
-	{
-		// [test case of data lineage]
-		// data1 (When there is no downstream for data to be tracked)
-		data1 := dummyData("data1", "run0")
-		t.Run(" Confirm that when only nodes exist in the graph, they can be output as dot format.", theory(
-			When{
-				Graph: lineage.DirectedGraph{
-					DataNodes:     map[string]lineage.DataNode{"data1": toDataNode(data1)},
-					RunNodes:      map[string]lineage.RunNode{},
-					EdgesFromData: map[string][]lineage.Edge{},
-					EdgesFromRun:  map[string][]lineage.Edge{},
-					RootNodes:     []string{},
-					KeysDataNode:  []string{"data1"},
-					KeysRunNode:   []string{},
-				},
-				ArgKnitId: "data1",
-			},
-			Then{
-				// The specification is to represent the node name as "r" + runId for run,
-				// and "d" + knitId for data.
-				RequiredContent: `digraph G {
-	node [shape=record fontsize=10]
-	edge [fontsize=10]
-
-	"ddata1"[
-		shape=none
-		color="#1c9930"
-		label=<
-			<TABLE CELLSPACING="0">
-				<TR><TD BGCOLOR="#1c9930"><FONT COLOR="#FFFFFF"><B>Data</B></FONT></TD><TD BGCOLOR="#d4ecc6">knit#id: data1</TD></TR>
-				<TR><TD COLSPAN="2"><FONT POINT-SIZE="8">2024-04-01T21:34:56+09:00</FONT></TD></TR>
-				<TR><TD COLSPAN="2"><B>foo</B>:bar<BR/><B>fizz</B>:bazz</TD></TR>
-			</TABLE>
-		>
-	];
-
-
-}`,
-			},
-		))
-	}
-	{
-		// [test case of data lineage]
-		// data1 --> [in/1] -->  run1 --> [out/1] --> data2
-		data1 := dummyData("data1", "run0", "run1")
-		data2 := dummyData("data2", "run1")
-		run1 := dummyRun("run1", map[string]string{"data1": "in/1"}, map[string]string{"data2": "out/1"})
-		t.Run(" Confirm that when only nodes, edges, exist in the graph, they can be output as dot format.", theory(
-			When{
-				Graph: lineage.DirectedGraph{
-					DataNodes:     map[string]lineage.DataNode{"data1": toDataNode(data1), "data2": toDataNode(data2)},
-					RunNodes:      map[string]lineage.RunNode{"run1": {run1.Summary}},
-					EdgesFromData: map[string][]lineage.Edge{"data1": {{"run1", "in/1"}}, "data2": {{"run1", "out/1"}}},
-					EdgesFromRun:  map[string][]lineage.Edge{"run1": {{"data1", "in/1"}, {"data2", "out/1"}}},
-					RootNodes:     []string{},
-					KeysDataNode:  []string{"data1", "data2"},
-					KeysRunNode:   []string{"run1"},
-				},
-				ArgKnitId: "data1",
-			},
-			Then{
-				RequiredContent: `digraph G {
-	node [shape=record fontsize=10]
-	edge [fontsize=10]
-
-	"ddata1"[
-		shape=none
-		color="#1c9930"
-		label=<
-			<TABLE CELLSPACING="0">
-				<TR><TD BGCOLOR="#1c9930"><FONT COLOR="#FFFFFF"><B>Data</B></FONT></TD><TD BGCOLOR="#d4ecc6">knit#id: data1</TD></TR>
-				<TR><TD COLSPAN="2"><FONT POINT-SIZE="8">2024-04-01T21:34:56+09:00</FONT></TD></TR>
-				<TR><TD COLSPAN="2"><B>foo</B>:bar<BR/><B>fizz</B>:bazz</TD></TR>
-			</TABLE>
-		>
-	];
-	"ddata2"[
-		shape=none
-		color="#1c9930"
-		label=<
-			<TABLE CELLSPACING="0">
-				<TR><TD BGCOLOR="#1c9930"><FONT COLOR="#FFFFFF"><B>Data</B></FONT></TD><TD BGCOLOR="#FFFFFF">knit#id: data2</TD></TR>
-				<TR><TD COLSPAN="2"><FONT POINT-SIZE="8">2024-04-01T21:34:56+09:00</FONT></TD></TR>
-				<TR><TD COLSPAN="2"><B>foo</B>:bar<BR/><B>fizz</B>:bazz</TD></TR>
-			</TABLE>
-		>
-	];
-	"rrun1"[
-		shape=none
-		color=orange
-		label=<
-			<TABLE CELLSPACING="0">
-				<TR><TD BGCOLOR="orange"><FONT COLOR="#FFFFFF"><B>Run</B></FONT></TD><TD><FONT COLOR="#007700"><B>done</B></FONT></TD><TD>id: run1</TD></TR>
-				<TR><TD COLSPAN="3"><FONT POINT-SIZE="8">last updated: 0001-01-01T09:18:59+09:18</FONT></TD></TR>
-				<TR><TD COLSPAN="3">image = test-image:test-version</TD></TR>
-			</TABLE>
-		>
-	];
-
-	"ddata1" -> "rrun1" [label="in/1"];
-	"ddata2" -> "rrun1" [label="out/1"];
-	"rrun1" -> "ddata1" [label="in/1"];
-	"rrun1" -> "ddata2" [label="out/1"];
-
-}`,
-			},
-		))
-	}
-	{
-		// [test case of data lineage]
-		// root -->  run1 --> [upload] --> data1 --> [in/1] --> run2 --> [out/1] --> data2
-		run1 := dummyRun("run1", map[string]string{}, map[string]string{"data1": "upload"})
-		data1 := dummyData("data1", "run1")
-		run2 := dummyRun("run2", map[string]string{"data1": "in/1"}, map[string]string{"data2": "out/1"})
-		data2 := dummyData("data2", "run2")
-		t.Run("Confirm that when nodes, edges, and roots exist in the graph, they can be output as dot format.", theory(
-			When{
-				Graph: lineage.DirectedGraph{
-					DataNodes:     map[string]lineage.DataNode{"data1": toDataNode(data1), "data2": toDataNode(data2)},
-					RunNodes:      map[string]lineage.RunNode{"run1": {run1.Summary}, "run2": {run2.Summary}},
-					EdgesFromData: map[string][]lineage.Edge{"data1": {{"run2", "in/1"}}, "data2": {}},
-					EdgesFromRun: map[string][]lineage.Edge{
-						"run1": {{ToId: "data1", Label: "upload"}}, "run2": {{ToId: "data2", Label: "out/1"}}},
-					KeysDataNode: []string{"data1", "data2"},
-					KeysRunNode:  []string{"run1", "run2"},
-					RootNodes:    []string{"run1"},
-				},
-				ArgKnitId: "data2",
-			},
-			Then{
-				RequiredContent: `digraph G {
-	node [shape=record fontsize=10]
-	edge [fontsize=10]
-
-	"ddata1"[
-		shape=none
-		color="#1c9930"
-		label=<
-			<TABLE CELLSPACING="0">
-				<TR><TD BGCOLOR="#1c9930"><FONT COLOR="#FFFFFF"><B>Data</B></FONT></TD><TD BGCOLOR="#FFFFFF">knit#id: data1</TD></TR>
-				<TR><TD COLSPAN="2"><FONT POINT-SIZE="8">2024-04-01T21:34:56+09:00</FONT></TD></TR>
-				<TR><TD COLSPAN="2"><B>foo</B>:bar<BR/><B>fizz</B>:bazz</TD></TR>
-			</TABLE>
-		>
-	];
-	"ddata2"[
-		shape=none
-		color="#1c9930"
-		label=<
-			<TABLE CELLSPACING="0">
-				<TR><TD BGCOLOR="#1c9930"><FONT COLOR="#FFFFFF"><B>Data</B></FONT></TD><TD BGCOLOR="#d4ecc6">knit#id: data2</TD></TR>
-				<TR><TD COLSPAN="2"><FONT POINT-SIZE="8">2024-04-01T21:34:56+09:00</FONT></TD></TR>
-				<TR><TD COLSPAN="2"><B>foo</B>:bar<BR/><B>fizz</B>:bazz</TD></TR>
-			</TABLE>
-		>
-	];
-	"rrun1"[
-		shape=none
-		color=orange
-		label=<
-			<TABLE CELLSPACING="0">
-				<TR><TD BGCOLOR="orange"><FONT COLOR="#FFFFFF"><B>Run</B></FONT></TD><TD><FONT COLOR="#007700"><B>done</B></FONT></TD><TD>id: run1</TD></TR>
-				<TR><TD COLSPAN="3"><FONT POINT-SIZE="8">last updated: 0001-01-01T09:18:59+09:18</FONT></TD></TR>
-				<TR><TD COLSPAN="3">image = test-image:test-version</TD></TR>
-			</TABLE>
-		>
-	];
-	"rrun2"[
-		shape=none
-		color=orange
-		label=<
-			<TABLE CELLSPACING="0">
-				<TR><TD BGCOLOR="orange"><FONT COLOR="#FFFFFF"><B>Run</B></FONT></TD><TD><FONT COLOR="#007700"><B>done</B></FONT></TD><TD>id: run2</TD></TR>
-				<TR><TD COLSPAN="3"><FONT POINT-SIZE="8">last updated: 0001-01-01T09:18:59+09:18</FONT></TD></TR>
-				<TR><TD COLSPAN="3">image = test-image:test-version</TD></TR>
-			</TABLE>
-		>
-	];
-	"root#0"[shape=Mdiamond];
-
-	"ddata1" -> "rrun2" [label="in/1"];
-	"rrun1" -> "ddata1" [label="upload"];
-	"rrun2" -> "ddata2" [label="out/1"];
-	"root#0" -> "rrun1";
-
-}`,
-			},
-		))
-	}
-	{
-		// [test case of data lineage]
-		// root -->  run1 --> [upload] --> data1 --> [in/1] --> run2 (failed) --> [out/1] --> data2
-		//                                                                                |-> log
-		run1 := dummyRun("run1", map[string]string{}, map[string]string{"data1": "upload"})
-		data1 := dummyData("data1", "run1")
-		log := dummyDataForFailed("log", "run2")
-		run2 := dummyFailedRunWithLog("run2", "log", map[string]string{"data1": "in/1"}, map[string]string{"data2": "out/1"})
-		data2 := dummyDataForFailed("data2", "run2")
-		t.Run("When there are failed run and its output, they can be output as dot format.", theory(
-			When{
-				Graph: lineage.DirectedGraph{
-					DataNodes:     map[string]lineage.DataNode{"data1": toDataNode(data1), "data2": toDataNode(data2), "log": toDataNode(log)},
-					RunNodes:      map[string]lineage.RunNode{"run1": {run1.Summary}, "run2": {run2.Summary}},
-					EdgesFromData: map[string][]lineage.Edge{"data1": {{"run2", "in/1"}}, "data2": {}, "log": {}},
-					EdgesFromRun: map[string][]lineage.Edge{
-						"run1": {{ToId: "data1", Label: "upload"}}, "run2": {{ToId: "data2", Label: "out/1"}, {ToId: "log", Label: "(log)"}}},
-					KeysDataNode: []string{"data1", "data2", "log"},
-					KeysRunNode:  []string{"run1", "run2"},
-					RootNodes:    []string{"run1"},
-				},
-				ArgKnitId: "data1",
-			},
-			Then{
-				RequiredContent: `digraph G {
-	node [shape=record fontsize=10]
-	edge [fontsize=10]
-
-	"ddata1"[
-		shape=none
-		color="#1c9930"
-		label=<
-			<TABLE CELLSPACING="0">
-				<TR><TD BGCOLOR="#1c9930"><FONT COLOR="#FFFFFF"><B>Data</B></FONT></TD><TD BGCOLOR="#d4ecc6">knit#id: data1</TD></TR>
-				<TR><TD COLSPAN="2"><FONT POINT-SIZE="8">2024-04-01T21:34:56+09:00</FONT></TD></TR>
-				<TR><TD COLSPAN="2"><B>foo</B>:bar<BR/><B>fizz</B>:bazz</TD></TR>
-			</TABLE>
-		>
-	];
-	"ddata2"[
-		shape=none
-		color="#1c9930"
-		label=<
-			<TABLE CELLSPACING="0">
-				<TR><TD BGCOLOR="#1c9930"><FONT COLOR="#FFFFFF"><B>Data</B></FONT></TD><TD BGCOLOR="#FFFFFF">knit#id: data2</TD></TR>
-				<TR><TD COLSPAN="2"><FONT POINT-SIZE="8">2024-04-01T21:34:56+09:00 | knit#transient:failed</FONT></TD></TR>
-				<TR><TD COLSPAN="2"><B>foo</B>:bar<BR/><B>fizz</B>:bazz</TD></TR>
-			</TABLE>
-		>
-	];
-	"dlog"[
-		shape=none
-		color="#1c9930"
-		label=<
-			<TABLE CELLSPACING="0">
-				<TR><TD BGCOLOR="#1c9930"><FONT COLOR="#FFFFFF"><B>Data</B></FONT></TD><TD BGCOLOR="#FFFFFF">knit#id: log</TD></TR>
-				<TR><TD COLSPAN="2"><FONT POINT-SIZE="8">2024-04-01T21:34:56+09:00 | knit#transient:failed</FONT></TD></TR>
-				<TR><TD COLSPAN="2"><B>foo</B>:bar<BR/><B>fizz</B>:bazz</TD></TR>
-			</TABLE>
-		>
-	];
-	"rrun1"[
-		shape=none
-		color=orange
-		label=<
-			<TABLE CELLSPACING="0">
-				<TR><TD BGCOLOR="orange"><FONT COLOR="#FFFFFF"><B>Run</B></FONT></TD><TD><FONT COLOR="#007700"><B>done</B></FONT></TD><TD>id: run1</TD></TR>
-				<TR><TD COLSPAN="3"><FONT POINT-SIZE="8">last updated: 0001-01-01T09:18:59+09:18</FONT></TD></TR>
-				<TR><TD COLSPAN="3">image = test-image:test-version</TD></TR>
-			</TABLE>
-		>
-	];
-	"rrun2"[
-		shape=none
-		color=orange
-		label=<
-			<TABLE CELLSPACING="0">
-				<TR><TD BGCOLOR="orange"><FONT COLOR="#FFFFFF"><B>Run</B></FONT></TD><TD><FONT COLOR="red"><B>failed</B></FONT></TD><TD>id: run2</TD></TR>
-				<TR><TD COLSPAN="3"><FONT POINT-SIZE="8">last updated: 0001-01-01T09:18:59+09:18</FONT></TD></TR>
-				<TR><TD COLSPAN="3">image = test-image:test-version</TD></TR>
-			</TABLE>
-		>
-	];
-	"root#0"[shape=Mdiamond];
-
-	"ddata1" -> "rrun2" [label="in/1"];
-	"rrun1" -> "ddata1" [label="upload"];
-	"rrun2" -> "ddata2" [label="out/1"];
-	"rrun2" -> "dlog" [label="(log)"];
-	"root#0" -> "rrun1";
-
-}`,
-			},
-		))
-	}
-	{
-		// [test case of data lineage]
-		//         	                                  data4 --> [in/3] -|
-		// data1 --> [in/1] -->  run1 --> [out/1] --> data2 --> [in/2] --> run2 --> [out/3] --> data5
-		//					          |-> [out/2] --> data3 --> [in/4] --> run3 --> [out/4] --> data6
-		data1 := dummyData("data1", "run0", "run1")
-		run1 := dummyRun("run1", map[string]string{"data1": "in/1"}, map[string]string{"data2": "out/1", "data3": "out/2"})
-		data2 := dummyData("data2", "run1", "run2")
-		data3 := dummyData("data3", "run1", "run3")
-		run2 := dummyRun("run2", map[string]string{"data2": "in/2", "data4": "in/3"}, map[string]string{"data5": "out/3"})
-		data4 := dummyData("data4", "runxx", "run2")
-		data5 := dummyData("data5", "run2")
-		run3 := dummyRun("run3", map[string]string{"data3": "in/4"}, map[string]string{"data6": "out/4"})
-		data6 := dummyData("data6", "run3")
-
-		t.Run("Confirm that when the graph configuration is complex, they can be output as dot format.", theory(
-			When{
-				Graph: lineage.DirectedGraph{
-					DataNodes: map[string]lineage.DataNode{
-						"data1": toDataNode(data1), "data2": toDataNode(data2), "data3": toDataNode(data3),
-						"data4": toDataNode(data4), "data5": toDataNode(data5), "data6": toDataNode(data6),
-					},
-					RunNodes: map[string]lineage.RunNode{
-						"run1": {run1.Summary}, "run2": {run2.Summary}, "run3": {run3.Summary},
-					},
-					EdgesFromData: map[string][]lineage.Edge{
-						"data1": {{"run1", "in/1"}}, "data2": {{"run2", "in/2"}}, "data3": {{"run3", "in/4"}},
-						"data4": {{"run2", "in/3"}}, "data5": {}, "data6": {},
-					},
-					EdgesFromRun: map[string][]lineage.Edge{
-						"run1": {{"data2", "out/1"}, {"data3", "out/2"}},
-						"run2": {{"data5", "out/3"}}, "run3": {{"data6", "out/4"}},
-					},
-					KeysDataNode: []string{"data1", "data2", "data3", "data4", "data5", "data6"},
-					KeysRunNode:  []string{"run1", "run2", "run3"},
-					RootNodes:    []string{},
-				},
-				ArgKnitId: "data1",
-			},
-			Then{
-				RequiredContent: `digraph G {
-	node [shape=record fontsize=10]
-	edge [fontsize=10]
-
-	"ddata1"[
-		shape=none
-		color="#1c9930"
-		label=<
-			<TABLE CELLSPACING="0">
-				<TR><TD BGCOLOR="#1c9930"><FONT COLOR="#FFFFFF"><B>Data</B></FONT></TD><TD BGCOLOR="#d4ecc6">knit#id: data1</TD></TR>
-				<TR><TD COLSPAN="2"><FONT POINT-SIZE="8">2024-04-01T21:34:56+09:00</FONT></TD></TR>
-				<TR><TD COLSPAN="2"><B>foo</B>:bar<BR/><B>fizz</B>:bazz</TD></TR>
-			</TABLE>
-		>
-	];
-	"ddata2"[
-		shape=none
-		color="#1c9930"
-		label=<
-			<TABLE CELLSPACING="0">
-				<TR><TD BGCOLOR="#1c9930"><FONT COLOR="#FFFFFF"><B>Data</B></FONT></TD><TD BGCOLOR="#FFFFFF">knit#id: data2</TD></TR>
-				<TR><TD COLSPAN="2"><FONT POINT-SIZE="8">2024-04-01T21:34:56+09:00</FONT></TD></TR>
-				<TR><TD COLSPAN="2"><B>foo</B>:bar<BR/><B>fizz</B>:bazz</TD></TR>
-			</TABLE>
-		>
-	];
-	"ddata3"[
-		shape=none
-		color="#1c9930"
-		label=<
-			<TABLE CELLSPACING="0">
-				<TR><TD BGCOLOR="#1c9930"><FONT COLOR="#FFFFFF"><B>Data</B></FONT></TD><TD BGCOLOR="#FFFFFF">knit#id: data3</TD></TR>
-				<TR><TD COLSPAN="2"><FONT POINT-SIZE="8">2024-04-01T21:34:56+09:00</FONT></TD></TR>
-				<TR><TD COLSPAN="2"><B>foo</B>:bar<BR/><B>fizz</B>:bazz</TD></TR>
-			</TABLE>
-		>
-	];
-	"ddata4"[
-		shape=none
-		color="#1c9930"
-		label=<
-			<TABLE CELLSPACING="0">
-				<TR><TD BGCOLOR="#1c9930"><FONT COLOR="#FFFFFF"><B>Data</B></FONT></TD><TD BGCOLOR="#FFFFFF">knit#id: data4</TD></TR>
-				<TR><TD COLSPAN="2"><FONT POINT-SIZE="8">2024-04-01T21:34:56+09:00</FONT></TD></TR>
-				<TR><TD COLSPAN="2"><B>foo</B>:bar<BR/><B>fizz</B>:bazz</TD></TR>
-			</TABLE>
-		>
-	];
-	"ddata5"[
-		shape=none
-		color="#1c9930"
-		label=<
-			<TABLE CELLSPACING="0">
-				<TR><TD BGCOLOR="#1c9930"><FONT COLOR="#FFFFFF"><B>Data</B></FONT></TD><TD BGCOLOR="#FFFFFF">knit#id: data5</TD></TR>
-				<TR><TD COLSPAN="2"><FONT POINT-SIZE="8">2024-04-01T21:34:56+09:00</FONT></TD></TR>
-				<TR><TD COLSPAN="2"><B>foo</B>:bar<BR/><B>fizz</B>:bazz</TD></TR>
-			</TABLE>
-		>
-	];
-	"ddata6"[
-		shape=none
-		color="#1c9930"
-		label=<
-			<TABLE CELLSPACING="0">
-				<TR><TD BGCOLOR="#1c9930"><FONT COLOR="#FFFFFF"><B>Data</B></FONT></TD><TD BGCOLOR="#FFFFFF">knit#id: data6</TD></TR>
-				<TR><TD COLSPAN="2"><FONT POINT-SIZE="8">2024-04-01T21:34:56+09:00</FONT></TD></TR>
-				<TR><TD COLSPAN="2"><B>foo</B>:bar<BR/><B>fizz</B>:bazz</TD></TR>
-			</TABLE>
-		>
-	];
-	"rrun1"[
-		shape=none
-		color=orange
-		label=<
-			<TABLE CELLSPACING="0">
-				<TR><TD BGCOLOR="orange"><FONT COLOR="#FFFFFF"><B>Run</B></FONT></TD><TD><FONT COLOR="#007700"><B>done</B></FONT></TD><TD>id: run1</TD></TR>
-				<TR><TD COLSPAN="3"><FONT POINT-SIZE="8">last updated: 0001-01-01T09:18:59+09:18</FONT></TD></TR>
-				<TR><TD COLSPAN="3">image = test-image:test-version</TD></TR>
-			</TABLE>
-		>
-	];
-	"rrun2"[
-		shape=none
-		color=orange
-		label=<
-			<TABLE CELLSPACING="0">
-				<TR><TD BGCOLOR="orange"><FONT COLOR="#FFFFFF"><B>Run</B></FONT></TD><TD><FONT COLOR="#007700"><B>done</B></FONT></TD><TD>id: run2</TD></TR>
-				<TR><TD COLSPAN="3"><FONT POINT-SIZE="8">last updated: 0001-01-01T09:18:59+09:18</FONT></TD></TR>
-				<TR><TD COLSPAN="3">image = test-image:test-version</TD></TR>
-			</TABLE>
-		>
-	];
-	"rrun3"[
-		shape=none
-		color=orange
-		label=<
-			<TABLE CELLSPACING="0">
-				<TR><TD BGCOLOR="orange"><FONT COLOR="#FFFFFF"><B>Run</B></FONT></TD><TD><FONT COLOR="#007700"><B>done</B></FONT></TD><TD>id: run3</TD></TR>
-				<TR><TD COLSPAN="3"><FONT POINT-SIZE="8">last updated: 0001-01-01T09:18:59+09:18</FONT></TD></TR>
-				<TR><TD COLSPAN="3">image = test-image:test-version</TD></TR>
-			</TABLE>
-		>
-	];
-
-	"ddata1" -> "rrun1" [label="in/1"];
-	"ddata2" -> "rrun2" [label="in/2"];
-	"ddata3" -> "rrun3" [label="in/4"];
-	"ddata4" -> "rrun2" [label="in/3"];
-	"rrun1" -> "ddata2" [label="out/1"];
-	"rrun1" -> "ddata3" [label="out/2"];
-	"rrun2" -> "ddata5" [label="out/3"];
-	"rrun3" -> "ddata6" [label="out/4"];
-
-}`,
-			},
-		))
-	}
 }
